Set the required trakt.tv API credentials

get_trakt_credentials searches for your credentials and stores them in the appropriate option variables. It also sets the HTTP header required for v2 API calls. To make this work automatically, place a key.json file either in the working directory or in ~/.config/trakt/key.json. Arguments to this function take precedence over any key file.

get_trakt_credentials(username = NULL, client.id = NULL,
  client.secret = NULL, set.headers = TRUE, silent = TRUE)

Arguments

username

Explicitly set your trakt.tv username

client.id

Explicitly set your APIv2 client id

client.secret

Explicitly set your APIv2 client secret (Not used yet)

set.headers

TRUE by default. Sets the httr headers for GET requests for the APIv2

silent

If TRUE (default), no messages are printed showing you the API information. Mostly for debug purposes.

Value

Nothing. Only messages.

Note

Please note that no oauth2 methods are supported yet, only client id really matters.

See also

Other API-basics: trakt.api.call, trakt.search.byid, trakt.search

Examples

## Not run: ------------------------------------
# # Use a key.json
# get_trakt_credentials()
# 
# # Explicitly set values
# get_trakt_credentials(username = "sean",
# client.id = "12fc1de7671c7f2fb4a8ac08ba7c9f45b447f4d5bad5e11e3490823d629afdf2")
## ---------------------------------------------