Get all movie / show aliases

Get all movie / show aliases

movies_aliases(id)

shows_aliases(id)

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things") for more details.

Source

movies_aliases() wraps endpoint movies/:id/aliases.

shows_aliases() wraps endpoint shows/:id/aliases.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

See also

Examples

movies_aliases(190430)

#> # A tibble: 16 x 2
#>    title                   country
#>    <chr>                   <chr>  
#>  1 X-Men Origins: Deadpool us     
#>  2 X-Men: Deadpool         us     
#>  3 Deadpool                br     
#>  4 Дедпул                  bg     
#>  5 Dedpul                  rs     
#>  6 Дэдпул                  ru     
#>  7 惡棍英雄：死侍          tw     
#>  8 死侍：不死现身          cn     
#>  9 X战警：死侍             cn     
#> 10 Deadpool 1              us     
#> 11 Masacre                 es     
#> 12 데드풀                  kr     
#> 13 死侍                    cn     
#> 14 08 Deadpool             cz     
#> 15 Deadpool 01 Deadpool    us     
#> 16 เดดพูล                   th     

shows_aliases(104439)

#> # A tibble: 43 x 2
#>    title                   country
#>    <chr>                   <chr>  
#>  1 Очень странные дела     us     
#>  2 Странные вещи           us     
#>  3 Загадкові справи        ua     
#>  4 Keisti dalykai          lt     
#>  5 Крайне странные события ru     
#>  6 უცნაური საქმეები        ge     
#>  7 Stranger Things (2016)  cn     
#>  8 Stranger Things         cn     
#>  9 Странные вещи           ru     
#> 10 Крайне странные события ru     
#> # … with 33 more rows