Show or movie user ratings

Returns a movie's or show's (or season's, or episode's) rating and ratings distribution. If you do not want the full ratings distribution, it is highly advised to just use *.summary functions or trakt.seasons.season for episode ratings.

trakt.shows.ratings(target)

trakt.movies.ratings(target)

trakt.seasons.ratings(target, season = 1L)

trakt.episodes.ratings(target, season = 1L, episode = 1L)

Arguments

target
character(1): The ID of the item requested. Preferably the Trakt ID (e.g. 1390), slug (e.g. "game-of-thrones"), or IMDb ID (e.g. "tt0944947"). Can also be of length greater than 1, in which case the function is called on all target values separately and the result is combined.

season
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

episode
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Value

A tibble().

Note

Since this function is able to work on multi-length inputs for target, season and episode, it is possible to get a lot of data, but at the cost of one API call per element in each argument. Please be kind to the API.

Examples

# A movie's ratings
trakt.movies.ratings("tron-legacy-2010")

#> # A tibble: 1 x 5
#>   rating votes distribution      id               type  
#>    <dbl> <int> <list>            <chr>            <chr> 
#> 1   7.27 12011 <tibble [10 × 2]> tron-legacy-2010 movies

# A show's ratings
trakt.shows.ratings("game-of-thrones")

#> # A tibble: 1 x 5
#>   rating votes distribution      id              type 
#>    <dbl> <int> <list>            <chr>           <chr>
#> 1   9.30 88132 <tibble [10 × 2]> game-of-thrones shows

# Ratings for seasons 1 through 5
trakt.seasons.ratings("futurama", season = 1:5)

#> # A tibble: 5 x 5
#>   rating votes distribution      id       season
#>    <dbl> <int> <list>            <chr>     <int>
#> 1   8.21   204 <tibble [10 × 2]> futurama      1
#> 2   8.33   151 <tibble [10 × 2]> futurama      2
#> 3   8.49   132 <tibble [10 × 2]> futurama      3
#> 4   8.26   124 <tibble [10 × 2]> futurama      4
#> 5   8.22   106 <tibble [10 × 2]> futurama      5

# Ratings for episodes 1 through 7 of season 1
trakt.episodes.ratings("futurama", season = 1, episode = 1:7)

#> # A tibble: 7 x 6
#>   rating votes distribution      id       season episode
#>    <dbl> <int> <list>            <chr>     <dbl>   <int>
#> 1   8.11  1909 <tibble [10 × 2]> futurama      1       1
#> 2   7.88  1493 <tibble [10 × 2]> futurama      1       2
#> 3   7.89  1361 <tibble [10 × 2]> futurama      1       3
#> 4   7.81  1213 <tibble [10 × 2]> futurama      1       4
#> 5   7.78  1167 <tibble [10 × 2]> futurama      1       5
#> 6   7.84  1150 <tibble [10 × 2]> futurama      1       6
#> 7   7.78  1068 <tibble [10 × 2]> futurama      1       7