Get a show or movie's (or season's or episode's) stats

The data contains watchers, playes, collectors, comments, lists, and votes.

shows_stats(id)

movies_stats(id)

seasons_stats(id, season = 1L)

episodes_stats(id, season = 1L, episode = 1L)

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things", package = "tRakt") for more details.

season
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

episode
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Source

shows_stats() wraps endpoint shows/:id/stats.

movies_stats() wraps endpoint movies/:id/stats.

seasons_stats() wraps endpoint shows/:id/seasons/:season/stats.

episodes_stats() wraps endpoint shows/:id/seasons/:season/episodes/:episode/stats.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

See also

Examples

# Stats for a movie
movies_stats("inception-2010")

#> # A tibble: 1 x 8
#>   watchers  plays collectors comments lists votes type   id            
#>      <int>  <int>      <int>    <int> <int> <int> <chr>  <chr>         
#> 1   311766 415689     217217      119 45394 54397 movies inception-2010

# \dontrun{
# Stats for multiple shows at once
shows_stats(c("breaking-bad", "game-of-thrones"))

#> # A tibble: 2 x 9
#>   watchers  plays collectors collected_episo… comments  lists  votes type  id   
#>      <int>  <int>      <int>            <int>    <int>  <int>  <int> <chr> <chr>
#> 1   548102 3.37e7     469198         23375650      436 138078  73437 shows brea…
#> 2  1113010 7.10e7    1388143         75954138      355 278641 105836 shows game…

# Stats for multiple episodes
episodes_stats("futurama", season = 1, episode = 1:7)

#> # A tibble: 7 x 9
#>   watchers  plays collectors comments lists votes season episode id      
#>      <int>  <int>      <int>    <int> <int> <int>  <dbl>   <int> <chr>   
#> 1   109959 146613     168093        2   181  2099      1       1 futurama
#> 2   103707 138277     167250        1   118  1637      1       2 futurama
#> 3   102456 136609     167679        0   107  1497      1       3 futurama
#> 4   100678 134238     167078        0   109  1342      1       4 futurama
#> 5    99637 132850     167291        1   109  1293      1       5 futurama
#> 6    99182 132243     166905        0   103  1265      1       6 futurama
#> 7    97798 130398     166694        0   102  1177      1       7 futurama

# }