Get show or movie summary info

Note that this function returns the same amount of informations as trakt.search.

trakt.movies.summary(target, extended = c("min", "full"))

trakt.shows.summary(target, extended = c("min", "full"))

Arguments

target
character(1): The ID of the item requested. Preferably the Trakt ID (e.g. 1390), slug (e.g. "game-of-thrones"), or IMDb ID (e.g. "tt0944947"). Can also be of length greater than 1, in which case the function is called on all target values separately and the result is combined.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

Value

A tibble().

Examples

# Minimal info by default
trakt.shows.summary("breaking-bad")

#> # A tibble: 1 x 7
#>   title         year trakt slug         tvdb  imdb      tmdb 
#>   <chr>        <int> <chr> <chr>        <chr> <chr>     <chr>
#> 1 Breaking Bad  2008 1388  breaking-bad 81189 tt0903747 1396 

# More information
trakt.shows.summary("breaking-bad", extended = "full")

#> # A tibble: 1 x 27
#>   title  year overview first_aired         runtime certification network country
#>   <chr> <int> <chr>    <dttm>                <int> <chr>         <chr>   <chr>  
#> 1 Brea…  2008 Walter … 2008-01-21 02:00:00      45 TV-MA         AMC     us     
#> # … with 19 more variables: trailer <chr>, homepage <chr>, status <chr>,
#> #   rating <dbl>, votes <int>, comment_count <int>, updated_at <dttm>,
#> #   language <chr>, available_translations <list>, genres <list>,
#> #   aired_episodes <int>, airs_day <chr>, airs_time <chr>, airs_timezone <chr>,
#> #   trakt <chr>, slug <chr>, tvdb <chr>, imdb <chr>, tmdb <chr>

# Info for multiple movies
trakt.movies.summary(c("inception-2010", "the-dark-knight-2008"), extended = "full")

#> # A tibble: 2 x 21
#>   title  year tagline overview released   runtime country trailer homepage
#>   <chr> <int> <chr>   <chr>    <date>       <int> <chr>   <chr>   <chr>   
#> 1 Ince…  2010 Your m… "Cobb, … 2010-07-16     148 us      http:/… http://…
#> 2 The …  2008 Why So… Batman … 2008-07-18     152 us      http:/… http://…
#> # … with 12 more variables: rating <dbl>, votes <int>, comment_count <int>,
#> #   updated_at <dttm>, language <chr>, available_translations <list>,
#> #   genres <list>, certification <chr>, trakt <chr>, slug <chr>, imdb <chr>,
#> #   tmdb <chr>