Get translations for a movie, show or episode

Get translations for a movie, show or episode

Usage

movies_translations(id, languages = NULL)

shows_translations(id, languages = NULL)

episodes_translations(id, season = 1L, episode = 1L, languages = NULL)

Source

movies_translations() wraps endpoint movies/:id/translations/:language.

shows_translations() wraps endpoint shows/:id/translations/:language.

episodes_translations() wraps endpoint shows/:id/seasons/:season/episodes/:episode/translations/:language.

Arguments

id

character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things") for more details.

languages

character(n): Two-letter language code(s). Also see trakt_languages for available languages (code and name).

season, episode

integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

See also

Examples

# Get all translations
movies_translations("193972")
#> # A tibble: 45 × 5
#>    title                                      overview  tagline language country
#>    <chr>                                      <chr>     <chr>   <chr>    <chr>  
#>  1 حكاية لعبة 4                               "استعد ل… مغامرة… ar       sa     
#>  2 Играта на играчките: Пътешествието         "Уди вин… Приклю… bg       bg     
#>  3 NA                                         "Compte!… Torna … ca       es     
#>  4 Toy Story 4: Příběh hraček                 "Kovboj … Připra… cs       cz     
#>  5 NA                                         "Siden A… NA      da       dk     
#>  6 Toy Story 4 - Alles hört auf kein Kommando "Die Cow… Hier s… de       de     
#>  7 Η Ιστορία των Παιχνιδιών 4                 "Στη διά… Η περι… el       gr     
#>  8 Toy Story 4                                "Woody h… Get Re… en       us     
#>  9 NA                                         "Cuando … Woody … es       mx     
#> 10 NA                                         "Woody s… La ave… es       es     
#> # ℹ 35 more rows

# Only get a specific language
movies_translations("193972", "de")
#> # A tibble: 1 × 5
#>   title                                      overview   tagline language country
#>   <chr>                                      <chr>      <chr>   <chr>    <chr>  
#> 1 Toy Story 4 - Alles hört auf kein Kommando Die Cowbo… Hier s… de       de