Get a single person's details

Get a single person's details, like their various IDs. If extended is "full", there will also be biographical data if available, e.g. their birthday.

Usage

people_summary(id, extended = c("min", "full"))

Source

people_summary() wraps endpoint people/:id.

Arguments

id

character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things") for more details.

extended

character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things") for more details.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

See also

Examples

# A single person's extended information
people_summary("bryan-cranston", "full")
#> # A tibble: 4 × 13
#>   name     socia…¹ biogr…² birthday   death birth…³ homep…⁴ known…⁵ gender trakt
#>   <chr>    <named> <chr>   <date>     <chr> <chr>   <chr>   <chr>   <chr>  <chr>
#> 1 Bryan C… <chr>   "Bryan… 1956-03-07 NA    San Fe… NA      acting  male   2977…
#> 2 Bryan C… <chr>   "Bryan… 1956-03-07 NA    San Fe… NA      acting  male   2977…
#> 3 Bryan C… <chr>   "Bryan… 1956-03-07 NA    San Fe… NA      acting  male   2977…
#> 4 Bryan C… <NULL>  "Bryan… 1956-03-07 NA    San Fe… NA      acting  male   2977…
#> # … with 3 more variables: slug <chr>, imdb <chr>, tmdb <chr>, and abbreviated
#> #   variable names ¹​social_ids, ²​biography, ³​birthplace, ⁴​homepage,
#> #   ⁵​known_for_department

# Multiple people
people_summary(c("kit-harington", "emilia-clarke"))
#> # A tibble: 2 × 5
#>   name          trakt  slug          imdb      tmdb   
#>   <chr>         <chr>  <chr>         <chr>     <chr>  
#> 1 Kit Harington 436512 kit-harington nm3229685 239019 
#> 2 Emilia Clarke 436511 emilia-clarke nm3592338 1223786