Result Filters

You can specifiy additional search filters in some functions. These filters are documented in the official API reference. For convenience, all supported filters are implemented as explicit, named arguments.

Arguments

query
character(1): Search titles and descriptions.

years
integer(1): 4-digit year or range, e.g. 2010-2020. Can also be an integer vector of length two which will be coerced appropriately.

genres
character(n): Genre slugs. Multiple values are allowed and will be concatenated.

languages
character(n): Two-digit language code. Also see languages for table of available languages.

countries
character(n): Two-character country code. See countries.

runtimes
character | integer: Numeric range in minutes, e.g. 30-90. Can also be an integer vector of length two which will be coerced appropriately.

ratings
character | integer: Numeric range between 0 and 100. Can also be an integer vector of length two which will be coerced appropriately.

certifications
character(n): Certification like pg-13. Multiple values are allowed. Use certifications for reference. Note that there are different certifications for shows and movies.

networks
Network name like HBO. See networks for a list of known networks.

status
character(n): The status of the shows. One of returning series, in production, planned, canceled, or ended