Get a shows next or latest episode

Get a shows next or latest episode

Usage

shows_next_episode(id, extended = c("min", "full"))

shows_last_episode(id, extended = c("min", "full"))

Source

shows_next_episode() wraps endpoint shows/:id/next_episode.

shows_last_episode() wraps endpoint shows/:id/last_episode.

Arguments

id

character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things") for more details.

extended

character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things") for more details.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

See also

Examples

shows_next_episode("one-piece")
#> # A tibble: 0 × 0
shows_last_episode("one-piece", extended = "full")
#> # A tibble: 1 × 17
#>   season number title             number_abs overview rating votes comment_count
#>    <int>  <int> <chr>                  <int> <chr>     <dbl> <int>         <int>
#> 1     22   1122 The Last Lesson!…       1122 Avalo P…   8.84   107             5
#> # ℹ 9 more variables: first_aired <dttm>, updated_at <dttm>,
#> #   available_translations <list>, runtime <int>, episode_type <chr>,
#> #   trakt <chr>, tvdb <chr>, imdb <chr>, tmdb <chr>