tRakt sitrep

Get general information about set options and APi credentaials.

Usage

tRakt_sitrep()

Details

Supported package-level options are:

tRakt_debug: logical(1): Should debug output be enabled? Default is FALSE.

tRakt_cache_dir: character(1): Where should the cache be stored?

tRakt_cache_max_age: numeric(1): How long should the cache be kept, in seconds? Default is one week (604800 seconds).

tRakt_cache_max_size: numeric(1): How big should the cache be, in bytes? Default is 100MB (100 * 1000^2 bytes).

All options can be overriden with environment variables of the same name.

Examples

tRakt_sitrep()
#> 
#> ── General settings ──
#> 
#> • Debug output enabled: FALSE
#> • Cache dir: "~/.cache/tRakt"
#> • Cache max age: 604800
#> • Cache max size: 1e+08
#> 
#> ── API authentication ──
#> 
#>   • Client ID:
#>   "12fc1de7671c7f2fb4a8ac08ba7c9f45b447f4d5bad5e11e3490823d629afdf2"
#>   • Can perform device authentication: FALSE