Get all the show data

trakt.get_full_showdata is a combination function of multiple functions in this package. The idea is to easily execute all major functions required to get a full show dataset.

trakt.get_full_showdata(query = NULL, slug = NULL, dropunaired = TRUE)

Arguments

query

Keyword used for trakt.search. Optional.

slug

Used if query is not specified. Optional, but gives exact results.

dropunaired

If TRUE, episodes which have not aired yet are dropped.

Value

A list containing multiple lists and data.frames with show info.

Note

This is primarily intended to be a convenience function for the case where you really want all that data. If you're just derping around, maybe you should consider interactively calling the other functions.

See also

Other show data: trakt.get_all_episodes, trakt.seasons.season, trakt.seasons.summary, trakt.show.people, trakt.show.ratings, trakt.show.summary, trakt.show.watching, trakt.shows.popular, trakt.shows.related, trakt.shows.trending, trakt.stats

Examples

## Not run: ------------------------------------
# get_trakt_credentials() # Set required API data/headers
# # Use the search within the function
# breakingbad <- trakt.get_full_showdata("Breaking Bad")
# # Alternatively, us a slug for explicit results
# breakingbad <- trakt.get_full_showdata(slug = "breaking-bad")
## ---------------------------------------------