Get a movie's release details

Retrieve one or more movie's release information, including the release date, country code (two letter, e.g. us), and the certification (e.g. PG).

trakt.movies.releases(target, country = NULL)

Arguments

target
character(1): The ID of the item requested. Preferably the Trakt ID (e.g. 1390), slug (e.g. "game-of-thrones"), or IMDb ID (e.g. "tt0944947"). Can also be of length greater than 1, in which case the function is called on all target values separately and the result is combined.

country
Optional two letter country code to filter by.

Value

A tibble().

Note

See the trakt API docs for further info

See also

Examples

trakt.movies.releases("tron-legacy-2010")

#> # A tibble: 19 x 6
#>    country certification release_date release_type note                  movie  
#>    <chr>   <chr>         <chr>        <chr>        <chr>                 <chr>  
#>  1 us      PG            2010-12-10   theatrical   World Premiere        tron-l…
#>  2 kz      Б14           2010-12-15   theatrical   <NA>                  tron-l…
#>  3 gr      PG            2010-12-16   theatrical   <NA>                  tron-l…
#>  4 au      PG            2010-12-16   theatrical   <NA>                  tron-l…
#>  5 us      PG            2010-12-17   theatrical   <NA>                  tron-l…
#>  6 gb      PG            2010-12-17   theatrical   <NA>                  tron-l…
#>  7 ca      ""            2010-12-17   theatrical   <NA>                  tron-l…
#>  8 se      11            2010-12-17   theatrical   <NA>                  tron-l…
#>  9 de      12            2010-12-17   theatrical   <NA>                  tron-l…
#> 10 br      12            2010-12-17   theatrical   <NA>                  tron-l…
#> 11 lt      ""            2010-12-24   theatrical   <NA>                  tron-l…
#> 12 it      ""            2010-12-29   theatrical   <NA>                  tron-l…
#> 13 kr      전체 관람가   2010-12-29   theatrical   <NA>                  tron-l…
#> 14 sk      12            2011-01-06   theatrical   <NA>                  tron-l…
#> 15 tr      PG-13         2011-01-28   theatrical   <NA>                  tron-l…
#> 16 nl      12            2011-01-28   theatrical   <NA>                  tron-l…
#> 17 fr      U             2011-02-09   theatrical   <NA>                  tron-l…
#> 18 us      PG            2011-04-05   physical     The Ultimate TRON Co… tron-l…
#> 19 se      11            2011-04-13   physical     <NA>                  tron-l…