Get trending movies

trakt.movies.trending returns a list of trending movies on trakt.tv. According to the API docs, it returns all movies being watched right now, where movies with the most users are returned first.

trakt.movies.trending(limit = 10, page = 1, extended = "min")

Arguments

limit

Number of movies to return. Is coerced to integer and must be greater than 0.

page

Page to return (default is 1) for http://docs.trakt.apiary.io/#introduction/pagination.

extended

Whether extended info should be provided. Defaults to "min", can either be "min" or "full"

Value

A data.frame containing trending movies with their number of watchers, name and ids

Note

See http://docs.trakt.apiary.io/reference/movies/trending/get-trending-movies

See also

Other aggregated data: trakt.movies.popular, trakt.movies.related, trakt.shows.popular, trakt.shows.related, trakt.shows.trending

Other movie data: trakt.movie.people, trakt.movie.ratings, trakt.movie.releases, trakt.movie.summary, trakt.movie.watching, trakt.movies.popular, trakt.movies.related, trakt.stats

Examples

## Not run: ------------------------------------
# get_trakt_credentials() # Set required API data/headers
# trakt.movies.trending(5)
## ---------------------------------------------