Get a single person's details

Get a single person's details, like their various IDs. If extended is "full", there will also be biographical data if available, e.g. their birthday.

trakt.people.summary(target, extended = c("min", "full"))

Arguments

target
character(1): The ID of the item requested. Preferably the Trakt ID (e.g. 1390), slug (e.g. "game-of-thrones"), or IMDb ID (e.g. "tt0944947"). Can also be of length greater than 1, in which case the function is called on all target values separately and the result is combined.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

Value

A tibble().

Details

This function wraps the API method /people/:id.

See also

Examples

# A single person's extended information
trakt.people.summary("bryan-cranston", "full")

#> # A tibble: 1 x 10
#>   name   biography  birthday   death birthplace homepage trakt slug  imdb  tmdb 
#>   <chr>  <chr>      <date>     <chr> <chr>      <chr>    <chr> <chr> <chr> <chr>
#> 1 Bryan… "An Ameri… 1956-03-07 <NA>  San Ferna… <NA>     2977… brya… nm01… 17419

# Multiple people
trakt.people.summary(c("kit-harington", "emilia-clarke"))

#> # A tibble: 2 x 5
#>   name          trakt  slug          imdb      tmdb   
#>   <chr>         <chr>  <chr>         <chr>     <chr>  
#> 1 Kit Harington 436512 kit-harington nm3229685 239019 
#> 2 Emilia Clarke 436511 emilia-clarke nm3592338 1223786