Search for a show via text query

trakt.search returns search result info.

trakt.search(query, type = "show", year = NULL)

Arguments

query

The keyword used for the search. If the query ends with a 4 digit numbers, this will be used as year parameter.

type

The type of data you're looking for. Defaults to show, can also be movie, episode, person or list.

year

Optionally filter by year.

Value

A data.frame containing a single search result. Hopefully the one you wanted. If no result is found, the return value is list(error = "Nothing found") and a warning

Details

Search for a show or movie with a keyword (e.g. "Breaking Bad") and receive basic info of the first search result. It's main use is to retrieve the ids or proper show/movie title for further use, as well as receiving a quick overview of a show/movie.

Note

See http://docs.trakt.apiary.io/reference/search/text-query

See also

Other API-basics: get_trakt_credentials, trakt.api.call, trakt.search.byid

Other search functions: trakt.search.byid

Examples

## Not run: ------------------------------------
# get_trakt_credentials() # Set required API data/headers
# breakingbad <- trakt.search("Breaking Bad")
## ---------------------------------------------