Get a season of a show

trakt.seasons.season gets a season's full data. Similar to trakt.seasons.summary, but this function returns full data for a single season, i.e. all the episodes of the season

trakt.seasons.season(target, seasons = 1L, extended = c("min", "full"))

Arguments

target
character(1): The ID of the item requested. Preferably the Trakt ID (e.g. 1390), slug (e.g. "game-of-thrones"), or IMDb ID (e.g. "tt0944947"). Can also be of length greater than 1, in which case the function is called on all target values separately and the result is combined.

seasons
integer(1) [1L]: The season(s) to get. Use 0 for special episodes.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

Value

A tibble().

Details

This function accesses this API method at the endpoint /shows/:show_id/seasons/:season_num.

Note

If you want to quickly gather episode data of all available seasons, see trakt.seasons.summary and use the episodes = TRUE parameter.

See also

Examples

trakt.seasons.season("breaking-bad", 1)

#> # A tibble: 7 x 7
#>   season episode title                         trakt tvdb   imdb      tmdb 
#>    <int>   <int> <chr>                         <chr> <chr>  <chr>     <chr>
#> 1      1       1 Pilot                         73482 349232 tt0959621 62085
#> 2      1       2 Cat's in the Bag...           73483 349233 tt1054724 62086
#> 3      1       3 ...And the Bag's in the River 73484 349235 tt1054725 62087
#> 4      1       4 Cancer Man                    73485 349236 tt1054726 62088
#> 5      1       5 Gray Matter                   73486 349238 tt1054727 62089
#> 6      1       6 Crazy Handful of Nothin'      73487 355100 tt1054728 62090
#> 7      1       7 A No-Rough-Stuff-Type Deal    73488 352534 tt1054729 62091