Search for related shows

trakt.shows.related returns shows related to the input show.

trakt.shows.related(target, extended = "min")

Arguments

target
The id of the movie requested. Either the slug (e.g. "tron-legacy-2010"), trakt id or IMDb id. If multiple targets are provided, the results will be rbinded together and a source column as appended, containing the provided id of the input.

extended
Whether extended info should be provided. Defaults to "min", can either be "min" or "full".

Value

A data.frame containing search results

Details

Receive a set of shows that are related to a specific show.

Note

See the trakt API docs for further info

See also

Examples

# NOT RUN {
get_trakt_credentials() # Set required API data/headers
related <- trakt.shows.related("game-of-thrones")
# }