Get a user's social connections

Retrieve a user's followers, followings or friends (the two-way relationship).

trakt.user.network(relationship = c("friends", "followers", "following"),
  user = getOption("trakt_username"), extended = c("min", "full"))

Arguments

relationship
character(1) ["friends"]: Type of user relationship. Either "friends", "followers", or "following".

user
Target username. Defaults to getOption("trakt_username"). Can also be of length greater than 1, in which case the function is called on all user values separately and the result is combined.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

Value

A tibble().

Details

The corresponding API methods are described in the API reference.

The relevant endpoints for this function are, depending on relationship:

followers: users/:id/followers

following: users/:id/following

friends: users/:id/friends

Note

Since no OAuth2 methods are supported yet, the specified user must not be private.

See also

Examples

# \dontrun{
trakt.user.network("friends", "jemus42")

#> # A tibble: 21 x 7
#>    friends_at          username       private name    vip   vip_ep slug         
#>    <dttm>              <chr>          <lgl>   <chr>   <lgl> <lgl>  <chr>        
#>  1 2012-08-03 19:15:59 sofakissen     FALSE   Daniel  FALSE FALSE  sofakissen   
#>  2 2012-08-03 21:52:54 Farthen        FALSE   <NA>    FALSE FALSE  farthen      
#>  3 2012-11-26 15:41:11 lsanoj         FALSE   Jonas   FALSE FALSE  lsanoj       
#>  4 2012-11-26 15:41:13 viirus         FALSE   <NA>    FALSE FALSE  viirus       
#>  5 2013-01-08 19:05:33 ekelias        FALSE   <NA>    FALSE FALSE  ekelias      
#>  6 2014-02-10 20:00:47 kimberley_jon… FALSE   ayn ra… FALSE FALSE  kimberley_jo…
#>  7 2014-02-14 21:47:33 Fenhl          FALSE   <NA>    FALSE FALSE  fenhl        
#>  8 2014-05-04 08:49:24 drseilzug      FALSE   <NA>    FALSE FALSE  drseilzug    
#>  9 2014-09-10 18:35:22 pwndbynck      FALSE   <NA>    FALSE FALSE  pwndbynck    
#> 10 2014-12-15 14:36:23 realtin        FALSE   <NA>    FALSE FALSE  realtin      
#> # … with 11 more rows

trakt.user.network("following", "jemus42")

#> # A tibble: 43 x 7
#>    followed_at         username      private name       vip   vip_ep slug       
#>    <dttm>              <chr>         <lgl>   <chr>      <lgl> <lgl>  <chr>      
#>  1 2012-08-03 19:15:59 sofakissen    FALSE   Daniel     FALSE FALSE  sofakissen 
#>  2 2012-08-03 21:52:54 Farthen       FALSE   <NA>       FALSE FALSE  farthen    
#>  3 2012-11-26 15:41:11 lsanoj        FALSE   Jonas      FALSE FALSE  lsanoj     
#>  4 2012-11-26 15:41:13 viirus        FALSE   <NA>       FALSE FALSE  viirus     
#>  5 2013-09-24 19:51:15 ekelias       FALSE   <NA>       FALSE FALSE  ekelias    
#>  6 2014-02-10 20:00:32 kimberley_jo… FALSE   ayn rant   FALSE FALSE  kimberley_…
#>  7 2014-02-14 21:48:30 Fenhl         FALSE   <NA>       FALSE FALSE  fenhl      
#>  8 2014-03-03 10:39:03 mxey          FALSE   <NA>       FALSE FALSE  mxey       
#>  9 2014-04-01 22:29:56 mmounirou     FALSE   " Pullo T… FALSE TRUE   mmounirou  
#> 10 2014-04-01 22:30:50 Draackje      FALSE   Draackje   FALSE FALSE  draackje   
#> # … with 33 more rows

# }