Get a user's ratings

Retrieve a user's media ratings

trakt.user.ratings(user = getOption("trakt.username"),
  type = c("movies", "seasons", "shows", "episodes"), rating = NULL,
  extended = c("min", "full"))

Arguments

user
Target username. Defaults to getOption("trakt.username"). Can also be of length greater than 1, in which case the function is called on all user values separately and the result is combined.

type
character(1): Either "shows" or "movies". For season/episode-specific functions, values seasons or episodes are also allowed.

rating
integer(1L): Rating to filter by. Can be 1 through 10, default is NULL

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

Value

A tibble().

See also

Examples

trakt.user.ratings(user = "jemus42")

#> # A tibble: 309 x 9
#>    rated_at            rating type  title      year trakt  slug      imdb  tmdb 
#>    <dttm>               <int> <chr> <chr>     <int> <chr>  <chr>     <chr> <chr>
#>  1 2019-03-07 22:47:20      7 movie Captain …  2019 193963 captain-… tt41… 2995…
#>  2 2019-02-21 20:06:20      7 movie The Girl…  2011 47598  the-girl… tt15… 65754
#>  3 2019-02-21 20:06:12      8 movie Serenity   2005 10041  serenity… tt03… 16320
#>  4 2019-02-21 20:03:12      6 movie Frozen     2013 77349  frozen-2… tt22… 1094…
#>  5 2019-02-15 10:53:58      7 movie Horns      2013 98504  horns-20… tt15… 1495…
#>  6 2019-01-22 22:58:29      6 movie Bird Box   2018 250049 bird-box… tt27… 4057…
#>  7 2018-12-30 15:26:48      7 movie Enemy      2013 114374 enemy-20… tt23… 1818…
#>  8 2018-12-29 06:06:38     10 movie Deadpool…  2018 226108 deadpool… tt54… 3834…
#>  9 2018-12-29 03:38:19      9 movie Sicario    2015 171369 sicario-… tt33… 2734…
#> 10 2018-12-29 03:38:16      8 movie The Acco…  2016 203309 the-acco… tt21… 3029…
#> # … with 299 more rows

trakt.user.ratings(user = "sean", type = "movies")

#> # A tibble: 144 x 9
#>    rated_at            rating type  title      year trakt  slug      imdb  tmdb 
#>    <dttm>               <int> <chr> <chr>     <int> <chr>  <chr>     <chr> <chr>
#>  1 2018-07-28 05:41:30      9 movie Watchmen   2009 7729   watchmen… tt04… 13183
#>  2 2018-02-12 20:22:17      9 movie Three Bi…  2017 266092 three-bi… tt50… 3599…
#>  3 2017-12-18 20:58:06      4 movie The Dark…  2017 222255 the-dark… tt16… 3534…
#>  4 2017-12-11 02:14:19      7 movie Kingsman…  2017 223444 kingsman… tt46… 3436…
#>  5 2017-12-11 02:14:09      9 movie American…  2017 260759 american… tt19… 4158…
#>  6 2017-11-16 03:29:00      5 movie Baywatch   2017 220314 baywatch… tt14… 3398…
#>  7 2017-08-02 21:25:33      8 movie Alien: C…  2017 86997  alien-co… tt23… 1268…
#>  8 2017-07-24 22:43:51      9 movie Baby Dri…  2017 219867 baby-dri… tt38… 3394…
#>  9 2017-07-21 04:43:03      8 movie Dunkirk    2017 224641 dunkirk-… tt50… 3747…
#> 10 2017-01-23 19:55:14      6 movie xXx: Ret…  2017 33214  xxx-retu… tt12… 47971
#> # … with 134 more rows