Get a user's watched shows or movies

trakt.user.watched retrieves a user's watched shows or movies. It does not use OAuth2, so you can only get data for a user with a public profile.

trakt.user.watched(user = getOption("trakt_username"),
  type = c("shows", "movies"), noseasons = TRUE)

Arguments

user
Target username. Defaults to getOption("trakt_username"). Can also be of length greater than 1, in which case the function is called on all user values separately and the result is combined.

type
character(1): Either "shows" or "movies". For season/episode-specific functions, values seasons or episodes are also allowed.

noseasons
logical(1) [TRUE]: Only for type = "show": Exclude detailed season data from output. This is advisable if you do not need per-episode data and want to be nice to the API.

Value

A tibble().

See also

Examples

# \dontrun{
# Use noseasons = TRUE to avoid receiving detailed season/episode data
trakt.user.watched(user = "sean", noseasons = TRUE)

#> Warning: Unknown or uninitialised column: 'first_aired'.

#> Warning: Unknown or uninitialised column: 'first_aired'.

#> # A tibble: 490 x 11
#>    plays last_watched_at     last_updated_at     reset_at title  year trakt
#>    <int> <dttm>              <dttm>              <lgl>    <chr> <int> <chr>
#>  1   167 2019-09-18 19:18:45 2019-09-18 19:18:45 NA       Marv…  2013 1394 
#>  2  1793 2019-09-18 14:50:00 2019-09-18 14:50:00 NA       Broo…  2013 48587
#>  3   802 2019-09-18 06:49:59 2019-09-18 06:49:59 NA       Rick…  2013 69829
#>  4     8 2019-09-18 04:23:51 2019-09-18 04:32:10 NA       Carn…  2019 1487…
#>  5   123 2019-09-17 22:59:21 2019-09-17 22:59:21 NA       Fast…  2012 46598
#>  6   568 2019-09-17 13:32:59 2019-09-17 13:32:59 NA       Work…  2011 36837
#>  7  3431 2019-09-17 13:26:57 2019-09-17 13:26:57 NA       How …  2005 1095 
#>  8     5 2019-09-16 21:36:26 2019-09-16 21:36:26 NA       On B…  2019 1335…
#>  9    43 2019-09-16 20:23:36 2019-09-16 20:23:36 NA       Ball…  2015 78111
#> 10  1760 2019-09-16 14:06:46 2019-09-16 14:06:46 NA       Comm…  2009 18265
#> # … with 480 more rows, and 4 more variables: slug <chr>, tvdb <chr>,
#> #   imdb <chr>, tmdb <chr>

# }