Get a user's watchlist

Retrieve a user's watchlisted shows or movies.

trakt.user.watchlist(user = getOption("trakt_username"),
  type = c("movies", "shows"), extended = c("min", "full"))

Arguments

user
character(1): Target username. Defaults to getOption("trakt_username"). Can also be of length greater than 1, in which case the function is called on all user values separately and the result is combined.

type
character(1): Either "shows" or "movies". For season/episode-specific functions, values seasons or episodes are also allowed.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

Value

A tibble().

Details

This function wraps the API method /users/:id/watchlist/:type.

See also

Examples

# \dontrun{
# Defaults to movie watchlist and minimal info
trakt.user.watchlist(user = "sean")

#> # A tibble: 328 x 10
#>     rank      id listed_at           type  title   year trakt slug   imdb  tmdb 
#>    <int>   <int> <dttm>              <chr> <chr>  <int> <chr> <chr>  <chr> <chr>
#>  1     1  7.27e7 2015-12-21 06:37:12 movie Creed   2015 2182… creed… tt30… 3122…
#>  2     2  7.36e7 2015-12-30 00:25:37 movie Strai…  2015 1749… strai… tt13… 2772…
#>  3     3  8.44e7 2016-03-04 03:31:33 movie Spotl…  2015 1941… spotl… tt18… 3143…
#>  4     4  8.44e7 2016-03-04 03:31:44 movie Room    2015 1633… room-… tt31… 2646…
#>  5     5  8.44e7 2016-03-04 07:04:18 movie Ghost…  2016 28880 ghost… tt12… 43074
#>  6     6  8.44e7 2016-03-04 07:06:25 movie Gambit  2020 1966… gambi… tt37… 3160…
#>  7     7  8.44e7 2016-03-04 07:10:34 movie High-…  2015 1547… high-… tt04… 2543…
#>  8     8  9.36e7 2016-05-11 19:56:23 movie Free …  2016 1970… free-… tt11… 3161…
#>  9     9  8.44e7 2016-03-04 07:18:16 movie Green…  2016 1934… green… tt40… 3139…
#> 10    10  9.36e7 2016-05-11 19:57:19 movie Money…  2016 2099… money… tt22… 3038…
#> # … with 318 more rows

# }