Get a user's social connections

Get followers, followings or friends (the two-way relationship).

Usage

user_followers(user = getOption("trakt_user"), extended = "min")

user_following(user = getOption("trakt_user"), extended = "min")

user_friends(user = getOption("trakt_user"), extended = "min")

Source

user_followers() wraps endpoint users/:id/followers.

user_following() wraps endpoint users/:id/following.

user_friends() wraps endpoint users/:id/friends.

Arguments

user

character(1): Target username (or slug). Defaults to "me", the OAuth user. Can also be of length greater than 1, in which case the function is called on all user values separately and the result is combined.

extended

character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things") for more details.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

Note

If the specified user is private, you need to be able to make an authenticated request and be friends with the user.

See also

Examples

if (FALSE) { # \dontrun{
user_followers(user = "sean")
} # }
if (FALSE) { # \dontrun{
user_following(user = "sean")
} # }
if (FALSE) { # \dontrun{
user_friends(user = "sean")
} # }