Returns stats about the movies, shows, and episodes a user has watched, collected, and rated.

Data about a user's interactions with movies, shows, seasons, episodes, as well as their social network (friends, followings, followers) and a frequency table of the user's media ratings so far.

Usage

user_stats(user = getOption("trakt_username"))

Source

user_stats() wraps endpoint users/:id/stats.

Arguments

user

character(1): Target username (or slug). Defaults to getOption("trakt_username"). Can also be of length greater than 1, in which case the function is called on all user values separately and the result is combined.

Value

A list of tibbles containing the following elements:

"movies"

"shows"

"seasons"

"episodes"

"network"

"ratings"

See also

Examples

if (FALSE) {
user_stats(user = "sean")
}