Most watched media

These functions return the most watched movies/shows on trakt.tv.

movies_watched(
  limit = 10,
  extended = c("min", "full"),
  period = c("weekly", "monthly", "yearly", "all"),
  query = NULL,
  years = NULL,
  genres = NULL,
  languages = NULL,
  countries = NULL,
  runtimes = NULL,
  ratings = NULL,
  certifications = NULL
)

shows_watched(
  limit = 10,
  extended = c("min", "full"),
  period = c("weekly", "monthly", "yearly", "all"),
  query = NULL,
  years = NULL,
  genres = NULL,
  languages = NULL,
  countries = NULL,
  runtimes = NULL,
  ratings = NULL,
  certifications = NULL,
  networks = NULL,
  status = NULL
)

Arguments

limit
integer(1) [10L]: Number of items to return. Must be greater than 0 and will be coerced via as.integer().

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things") for more details.

period
character(1) ["weekly"]: Which period to filter by. Possible values are "weekly", "monthly", "yearly", "all".

query
character(1): Search string for titles and descriptions. For search_query() other fields are searched depending on the type of media. See the API docs for a full reference.

years
character | integer: 4-digit year (2010) or range, e.g. "2010-2020". Can also be an integer vector of length two which will be coerced appropriately, e.g. c(2010, 2020).

genres
character(n): Genre slug(s). See trakt_genres for a table of genres. Multiple values are allowed and will be concatenated.

languages
character(n): Two-letter language code(s). Also see trakt_languages for available languages (code and name).

countries
character(n): Two-letter country code(s). See trakt_countries.

runtimes
character | integer: Integer range in minutes, e.g. 30-90. Can also be an integer vector of length two which will be coerced appropriately.

ratings
character | integer: Integer range between 0 and 100. Can also be an integer vector of length two which will be coerced appropriately. Note that user-supplied ratings are in the range of 1 to 10, yet the ratings on the site itself are scaled to the range of 1 to 100.

certifications
character(n): Certification(s) like pg-13. Multiple values are allowed. Use trakt_certifications for reference. Note that there are different certifications for shows and movies.

networks
character(n): (Shows only) Network name like HBO. See trakt_networks for a list of known networks.

status
character(n): (Shows only) The status of the shows. One of "returning series", "in production", "planned", "canceled", or "ended".

Source

movies_watched() wraps endpoint movies/watched/:period.

shows_watched() wraps endpoint shows/watched/:period.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

The Dynamic Lists on trakt.tv

These functions access the automatically updated lists provided by trakt.tv. Each function comes in two flavors: Shows or movies. The following descriptions are adapted directly from the API reference.

Popular: Popularity is calculated using the rating percentage and the number of ratings.

Trending: Returns all movies/shows being watched right now. Movies/shows with the most users are returned first.

Played: Returns the most played (a single user can watch multiple times) movies/shows in the specified time period.

Watched: Returns the most watched (unique users) movies/shows in the specified time period.

Collected: Returns the most collected (unique users) movies/shows in the specified time period.

Anticipated: Returns the most anticipated movies/shows based on the number of lists a movie/show appears on.

Updates: Returns all movies/shows updated since the specified UTC start_date. In this case, the upper bound for limit is 100.

The functions for Played, Watched, Collected and Played each return the same additional variables besides the media information: watcher_count, play_count, collected_count, collector_count.

See also