Anticipated media

These functions return the most anticipated movies/shows on trakt.tv.

movies_anticipated(limit = 10, extended = c("min", "full"),
  query = NULL, years = NULL, genres = NULL, languages = NULL,
  countries = NULL, runtimes = NULL, ratings = NULL,
  certifications = NULL, networks = NULL, status = NULL)

shows_anticipated(limit = 10, extended = c("min", "full"),
  query = NULL, years = NULL, genres = NULL, languages = NULL,
  countries = NULL, runtimes = NULL, ratings = NULL,
  certifications = NULL, networks = NULL, status = NULL)

Arguments

limit
integer(1) [10L]: Number of items to return. Must be greater than 0 and will be coerced to integer.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See section "Extended Information" for more details.

query
character(1): Search string for titles and descriptions. For search_query() other fields are searched depending on the type of media. See the API docs for a full reference.

years
character | integer: 4-digit year (2010) or range, e.g. "2010-2020". Can also be an integer vector of length two which will be coerced appropriately.

genres
character(n): Genre slug(s). See trakt_genres for a table of genres. Multiple values are allowed and will be concatenated.

languages
character(n): Two-digit language code(s). Also see trakt_languages for table of available languages.

countries
character(n): Two-character country code(s). See trakt_countries.

runtimes
character | integer: Integer range in minutes, e.g. 30-90. Can also be an integer vector of length two which will be coerced appropriately.

ratings
character | integer: Integer range between 0 and 100. Can also be an integer vector of length two which will be coerced appropriately.

certifications
character(n): Certification(s) like pg-13. Multiple values are allowed. Use trakt_certifications for reference. Note that there are different certifications for shows and movies.

networks
character(n): (Shows only) Network name like HBO. See trakt_networks for a list of known networks.

status
character(n): (Shows only) The status of the shows. One of "returning series", "in production", "planned", "canceled", or "ended".

The Dynamic Lists on trakt.tv

These functions access the automatically updated lists provided by trakt.tv. Each function comes in two flavors: Shows or movies. The following descriptions are adapted directly from the API reference.

Popular: Popularity is calculated using the rating percentage and the number of ratings.

Trending: Returns all movies/shows being watched right now. Movies/shows with the most users are returned first.

Played: Returns the most played (a single user can watch multiple times) movies/shows in the specified time period.

Watched: Returns the most watched (unique users) movies/shows in the specified time period.

Collected: Returns the most collected (unique users) movies/shows in the specified time period.

Anticipated: Returns the most anticipated movies/shows based on the number of lists a movie/show appears on.

Updates: Returns all movies/shows updated since the specified UTC start_date. In this case, the upper bound for limit is 100.

The functions for Played, Watched, Collected and Played each return the same additional variables besides the media information: watcher_count, play_count, collected_count, collector_count.

See also

Examples

# \dontrun{
# Get 15 the most anticipated upcoming shows on Netflix that air this year
current_year <- format(Sys.Date(), "%Y")
shows_anticipated(limit = 15, networks = "Netflix", years = current_year)

#> # A tibble: 15 x 8
#>    list_count title              year trakt  slug            tvdb  imdb    tmdb 
#>         <int> <chr>             <int> <chr>  <chr>           <chr> <chr>   <chr>
#>  1       9643 The Witcher        2019 138163 the-witcher-20… 3626… tt5180… 71912
#>  2       1279 Living with Your…  2019 151776 living-with-yo… 3685… tt8880… 81532
#>  3        705 Daybreak           2019 152108 daybreak-2019   3547… tt8755… 83478
#>  4        296 V-Wars             2019 133365 v-wars          3493… tt7403… <NA> 
#>  5        170 The Faction        2019 141592 the-faction     3566… <NA>    <NA> 
#>  6        139 Sintonia           2019 131399 sintonia        3443… tt8165… <NA> 
#>  7         97 Rhythm + Flow      2019 152211 rhythm-flow     3690… tt9278… 83996
#>  8         56 Green Eggs and H…  2019 131694 green-eggs-and… 3486… tt4651… 86957
#>  9         36 Nowhere Man        2019 153240 nowhere-man-15… 3700… tt1089… <NA> 
#> 10         29 Hache              2019 147926 hache           3555… tt8575… 84112
#> 11         19 Cidades Invisive…  2019 134761 cidades-invisi… 3493… <NA>    <NA> 
#> 12          9 Unnatural Select…  2019 153627 unnatural-sele… 3704… tt1106… <NA> 
#> 13          7 Brotherhood        2019 153095 brotherhood-20… 3589… tt8905… <NA> 
#> 14          7 Queer Eye: We're…  2019 152948 queer-eye-we-r… 3697… <NA>    <NA> 
#> 15          7 Prank Encounters   2019 152904 prank-encounte… 3697… <NA>    <NA> 

# }