See which movies / shows are popular

According to the API docs, popularity is calculated based both ratings and number of ratings. Trending items are those being watched right now, where items with the most users watching are returned first. Anticipation is measured by the number of user-created lists an items is part of while not being released yet.

trakt.popular(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"))

trakt.trending(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"))

trakt.anticipated(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"))

trakt.played(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"))

trakt.watched(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"))

trakt.collected(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"))

trakt.updates(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), start_date = Sys.Date() - 7)

Arguments

type
character(1): Either "shows" or "movies". For season/episode-specific functions, values seasons or episodes are also allowed.

limit
integer(1) [10L]: Number of items to return. Must be greater than 0 and will be coerced to integer.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

period
character(1) ["weekly"]: Which period to filter by. Possible values are "weekly", "monthly", "yearly", "all".

start_date
character(1): A date in the past from which on to count updates. If no date is supplied, the default is to use the date 7 days in the past relative to the current date. Value must either be standard YYYY-MM-DD format or an object of class Date, which will then be coerced via as.character().

Value

A tibble().

Details

These functions access the automatically updated lists provided by trakt.tv. Each function comes in two flavors: Shows or movies. The following descriptions are adapted directly from the API reference.

Popular: Popularity is calculated using the rating percentage and the number of ratings.

Trending: Returns all movies/shows being watched right now. Movies/shows with the most users are returned first.

Played: Returns the most played (a single user can watch multiple times) movies/shows in the specified time period.

Watched: Returns the most watched (unique users) movies/shows in the specified time period.

Collected: Returns the most collected (unique users) movies/shows in the specified time period.

Anticipated: Returns the most anticipated movies/shows based on the number of lists a movie/show appears on.

Updates: Returns all movies/shows updated since the specified UTC start_date. In this case, the upper bound for limit is 100.

The functions for Played, Watched, Collected and Played each return the same additional variables besides the media information: watcher_count, play_count, collected_count, collector_count.

Examples

# Get popular shows with only ids
trakt.popular(type = "shows")

#> # A tibble: 10 x 7
#>    title                  year trakt slug                  tvdb   imdb     tmdb 
#>    <chr>                 <int> <chr> <chr>                 <chr>  <chr>    <chr>
#>  1 Game of Thrones        2011 1390  game-of-thrones       121361 tt09449… 1399 
#>  2 Breaking Bad           2008 1388  breaking-bad          81189  tt09037… 1396 
#>  3 The Walking Dead       2010 1393  the-walking-dead      153021 tt15202… 1402 
#>  4 The Big Bang Theory    2007 1409  the-big-bang-theory   80379  tt08982… 1418 
#>  5 Sherlock               2010 19792 sherlock              176941 tt14755… 19885
#>  6 How I Met Your Mother  2005 1095  how-i-met-your-mother 75760  tt04606… 1100 
#>  7 Dexter                 2006 1396  dexter                79349  tt07732… 1405 
#>  8 Arrow                  2012 1403  arrow                 257655 tt21930… 1412 
#>  9 Friends                1994 1657  friends               79168  tt01087… 1668 
#> 10 Lost                   2004 4583  lost-2004             73739  tt04110… 4607 

# Get trending movies with extended information
trakt.trending(type = "movies", 5, extended = "full")

#> # A tibble: 5 x 22
#>   watchers title  year tagline overview released   runtime country trailer
#>      <int> <chr> <int> <chr>   <chr>    <date>       <int> <chr>   <chr>  
#> 1      102 Spid…  2018 More T… "Miles … 2018-12-14     117 us      http:/…
#> 2       68 Bohe…  2018 Fearle… Singer … 2018-11-02     135 us      http:/…
#> 3       67 Gree…  2018 Inspir… Tony Li… 2018-11-21     130 us      http:/…
#> 4       67 Inst…  2018 Just a… When Pe… 2018-11-16     118 us      http:/…
#> 5       50 Aqua…  2018 Home I… Once ho… 2018-12-21     143 us      http:/…
#> # … with 13 more variables: homepage <chr>, rating <dbl>, votes <int>,
#> #   comment_count <int>, updated_at <dttm>, language <chr>,
#> #   available_translations <list>, genres <list>, certification <chr>,
#> #   trakt <int>, slug <chr>, imdb <chr>, tmdb <int>

# Get top 5 anticipated movies
trakt.anticipated(type = "movies", 5)

#> # A tibble: 5 x 7
#>   list_count title               year  trakt slug                 imdb      tmdb
#>        <int> <chr>              <int>  <int> <chr>                <chr>    <int>
#> 1      33994 Captain Marvel      2019 193963 captain-marvel-2019  tt4154… 299537
#> 2      33513 Avengers: Endgame   2019 191798 avengers-endgame-20… tt4154… 299534
#> 3      28479 Avatar: The Way o…  2020  56580 avatar-the-way-of-w… tt1630…  76600
#> 4      26084 Star Wars: Episod…  2019 114335 star-wars-episode-i… tt2527… 181812
#> 5      20907 The Flash           2020 195144 the-flash-2020       tt0439… 298618