See which movies / shows are popular

According to the API docs, popularity is calculated based both ratings and number of ratings. Trending items are those being watched right now, where items with the most users watching are returned first. Anticipation is measured by the number of user-created lists an items is part of while not being released yet.

trakt.popular(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), query = NULL, years = NULL,
  genres = NULL, languages = NULL, countries = NULL,
  runtimes = NULL, ratings = NULL, certifications = NULL,
  networks = NULL, status = NULL)

trakt.trending(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), query = NULL, years = NULL,
  genres = NULL, languages = NULL, countries = NULL,
  runtimes = NULL, ratings = NULL, certifications = NULL,
  networks = NULL, status = NULL)

trakt.anticipated(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), query = NULL, years = NULL,
  genres = NULL, languages = NULL, countries = NULL,
  runtimes = NULL, ratings = NULL, certifications = NULL,
  networks = NULL, status = NULL)

trakt.played(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"), query = NULL, years = NULL, genres = NULL,
  languages = NULL, countries = NULL, runtimes = NULL,
  ratings = NULL, certifications = NULL, networks = NULL,
  status = NULL)

trakt.watched(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"), query = NULL, years = NULL, genres = NULL,
  languages = NULL, countries = NULL, runtimes = NULL,
  ratings = NULL, certifications = NULL, networks = NULL,
  status = NULL)

trakt.collected(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"), query = NULL, years = NULL, genres = NULL,
  languages = NULL, countries = NULL, runtimes = NULL,
  ratings = NULL, certifications = NULL, networks = NULL,
  status = NULL)

trakt.updates(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), start_date = Sys.Date() - 7)

Arguments

type
character(1): Either "shows" or "movies". For season/episode-specific functions, values seasons or episodes are also allowed.

limit
integer(1) [10L]: Number of items to return. Must be greater than 0 and will be coerced to integer.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

query
character(1): Search titles and descriptions.

years
integer(1): 4-digit year or range, e.g. 2010-2020. Can also be an integer vector of length two which will be coerced appropriately.

genres
character(n): Genre slug(s). See genres for a table of genres. Multiple values are allowed and will be concatenated.

languages
character(n): Two-digit language code(s). Also see languages for table of available languages.

countries
character(n): Two-character country code(s). See countries.

runtimes
character | integer: Integer range in minutes, e.g. 30-90. Can also be an integer vector of length two which will be coerced appropriately.

ratings
character | integer: Integer range between 0 and 100. Can also be an integer vector of length two which will be coerced appropriately.

certifications
character(n): Certification(s) like pg-13. Multiple values are allowed. Use certifications for reference. Note that there are different certifications for shows and movies.

networks
character(n): (Shows only) Network name like HBO. See networks for a list of known networks.

status
character(n): (Shows only) The status of the shows. One of returning series, in production, planned, canceled, or ended

period
character(1) ["weekly"]: Which period to filter by. Possible values are "weekly", "monthly", "yearly", "all".

start_date
character(1): A date in the past from which on to count updates. If no date is supplied, the default is to use the date 7 days in the past relative to the current date. Value must either be standard YYYY-MM-DD format or an object of class Date, which will then be coerced via as.character().

Value

A tibble().

Details

These functions access the automatically updated lists provided by trakt.tv. Each function comes in two flavors: Shows or movies. The following descriptions are adapted directly from the API reference.

Popular: Popularity is calculated using the rating percentage and the number of ratings.

Trending: Returns all movies/shows being watched right now. Movies/shows with the most users are returned first.

Played: Returns the most played (a single user can watch multiple times) movies/shows in the specified time period.

Watched: Returns the most watched (unique users) movies/shows in the specified time period.

Collected: Returns the most collected (unique users) movies/shows in the specified time period.

Anticipated: Returns the most anticipated movies/shows based on the number of lists a movie/show appears on.

Updates: Returns all movies/shows updated since the specified UTC start_date. In this case, the upper bound for limit is 100.

The functions for Played, Watched, Collected and Played each return the same additional variables besides the media information: watcher_count, play_count, collected_count, collector_count.

Note

trakt.updates does not support filters.

See also

Examples

# Get popular shows with only ids
trakt.popular(type = "shows")

#> # A tibble: 10 x 7
#>    title                  year trakt slug                  tvdb   imdb     tmdb 
#>    <chr>                 <int> <chr> <chr>                 <chr>  <chr>    <chr>
#>  1 Game of Thrones        2011 1390  game-of-thrones       121361 tt09449… 1399 
#>  2 Breaking Bad           2008 1388  breaking-bad          81189  tt09037… 1396 
#>  3 The Walking Dead       2010 1393  the-walking-dead      153021 tt15202… 1402 
#>  4 The Big Bang Theory    2007 1409  the-big-bang-theory   80379  tt08982… 1418 
#>  5 Sherlock               2010 19792 sherlock              176941 tt14755… 19885
#>  6 How I Met Your Mother  2005 1095  how-i-met-your-mother 75760  tt04606… 1100 
#>  7 Dexter                 2006 1396  dexter                79349  tt07732… 1405 
#>  8 Arrow                  2012 1403  arrow                 257655 tt21930… 1412 
#>  9 Friends                1994 1657  friends               79168  tt01087… 1668 
#> 10 Lost                   2004 4583  lost-2004             73739  tt04110… 4607 

# Get trending movies with extended information
trakt.trending(type = "movies", 5, extended = "full")

#> # A tibble: 5 x 22
#>   watchers title  year tagline overview released   runtime country trailer
#>      <int> <chr> <int> <chr>   <chr>    <date>       <int> <chr>   <chr>  
#> 1      135 Aqua…  2018 Home I… Once ho… 2018-12-21     143 us      http:/…
#> 2      121 Capt…  2019 Higher… The sto… 2019-03-08     124 us      http:/…
#> 3       84 Spid…  2018 More T… "Miles … 2018-12-14     117 us      http:/…
#> 4       74 Mary…  2018 Magic … In Depr… 2018-12-19     131 us      http:/…
#> 5       45 Ralp…  2018 Ralph … "Six ye… 2018-11-20     112 us      http:/…
#> # … with 13 more variables: homepage <chr>, rating <dbl>, votes <int>,
#> #   comment_count <int>, updated_at <dttm>, language <chr>,
#> #   available_translations <list>, genres <list>, certification <chr>,
#> #   trakt <int>, slug <chr>, imdb <chr>, tmdb <int>

# Get top 5 anticipated movies
trakt.anticipated(type = "movies", 5)

#> # A tibble: 5 x 7
#>   list_count title               year  trakt slug                 imdb      tmdb
#>        <int> <chr>              <int>  <int> <chr>                <chr>    <int>
#> 1      33994 Captain Marvel      2019 193963 captain-marvel-2019  tt4154… 299537
#> 2      33513 Avengers: Endgame   2019 191798 avengers-endgame-20… tt4154… 299534
#> 3      28479 Avatar: The Way o…  2020  56580 avatar-the-way-of-w… tt1630…  76600
#> 4      26084 Star Wars: Episod…  2019 114335 star-wars-episode-i… tt2527… 181812
#> 5      20907 The Flash           2020 195144 the-flash-2020       tt0439… 298618

# Get 15 the most anticipated upcoming shows on Netflix that air this year
current_year <- format(Sys.Date(), "%Y")
trakt.anticipated("shows", limit = 15, networks = "Netflix", years = current_year)

#> # A tibble: 15 x 8
#>    list_count title             year trakt  slug             tvdb  imdb    tmdb 
#>         <int> <chr>            <int> <chr>  <chr>            <chr> <chr>   <chr>
#>  1        910 Ultraman          2019 133835 ultraman-2019    3499… <NA>    80795
#>  2        657 After Life        2019 130511 after-life       3475… tt8398… <NA> 
#>  3        628 Love, Death & R…  2019 142611 love-death-robo… 3578… tt9561… <NA> 
#>  4        323 Turn Up Charlie   2019 135212 turn-up-charlie  3515… tt8304… 85486
#>  5        282 Our Planet        2019 139672 our-planet-2019  3552… tt9253… 83880
#>  6        200 Osmosis           2019 143847 osmosis-2019     3592… tt7043… 85424
#>  7        168 Formula 1 Drive…  2019 144377 formula-1-drive… 3599… tt8289… <NA> 
#>  8        129 Quicksand         2019 143600 quicksand        3589… tt8686… 86621
#>  9         89 7 Seeds           2019 140359 7-seeds          3560… tt9348… 85940
#> 10         46 Most Beautiful …  2019 133896 most-beautiful-… 3374… tt8001… <NA> 
#> 11         17 ULTRAMAN          2019 140389 ultraman-2019-1… 3560… <NA>    80795
#> 12         16 Shadow            2019 145021 shadow-2019      3606… tt9817… <NA> 
#> 13          9 Sintonia          2019 131399 sintonia         3443… tt8165… <NA> 
#> 14          8 When They See Us  2019 144923 when-they-see-us 3603… tt7137… 81355
#> 15          3 No Good Nick      2019 144963 no-good-nick-20… 3606… <NA>    <NA> 

# Get the most popular German-language movies between 1990 and 2010
trakt.popular("movies", languages = "de", years = c(1990, 2010))

#> # A tibble: 10 x 6
#>    title                    year trakt slug                       imdb      tmdb
#>    <chr>                   <int> <int> <chr>                      <chr>    <int>
#>  1 Downfall                 2004   489 downfall-2004              tt03631…   613
#>  2 The Lives of Others      2006   461 the-lives-of-others-2006   tt04050…   582
#>  3 Run Lola Run             1998    73 run-lola-run-1998          tt01308…   104
#>  4 The Wave                 2008  3954 the-wave-2008              tt10636…  7735
#>  5 Good bye, Lenin!         2003   285 good-bye-lenin-2003        tt03013…   338
#>  6 The Experiment           2001   454 the-experiment-2001        tt02502…   575
#>  7 Funny Games              1997  5445 funny-games-1997           tt01191… 10234
#>  8 The White Ribbon         2009 24371 the-white-ribbon-2009      tt11493… 37903
#>  9 The Baader Meinhof Com…  2008  3789 the-baader-meinhof-comple… tt07654…  6968
#> 10 Cargo                    2009 21807 cargo-2009                 tt03819… 34069