See which movies / shows are popular

According to the API docs, popularity is calculated based both on ratings and the number of ratings (i.e., votes). Trending items are those being watched right now, where items with the most users currently watching are returned first. Anticipation is measured by the number of user-created lists an items is part of while not being released yet.

trakt.popular(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), query = NULL, years = NULL,
  genres = NULL, languages = NULL, countries = NULL,
  runtimes = NULL, ratings = NULL, certifications = NULL,
  networks = NULL, status = NULL)

trakt.trending(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), query = NULL, years = NULL,
  genres = NULL, languages = NULL, countries = NULL,
  runtimes = NULL, ratings = NULL, certifications = NULL,
  networks = NULL, status = NULL)

trakt.anticipated(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), query = NULL, years = NULL,
  genres = NULL, languages = NULL, countries = NULL,
  runtimes = NULL, ratings = NULL, certifications = NULL,
  networks = NULL, status = NULL)

trakt.played(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"), query = NULL, years = NULL, genres = NULL,
  languages = NULL, countries = NULL, runtimes = NULL,
  ratings = NULL, certifications = NULL, networks = NULL,
  status = NULL)

trakt.watched(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"), query = NULL, years = NULL, genres = NULL,
  languages = NULL, countries = NULL, runtimes = NULL,
  ratings = NULL, certifications = NULL, networks = NULL,
  status = NULL)

trakt.collected(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"), query = NULL, years = NULL, genres = NULL,
  languages = NULL, countries = NULL, runtimes = NULL,
  ratings = NULL, certifications = NULL, networks = NULL,
  status = NULL)

trakt.updates(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), start_date = Sys.Date() - 7)

Arguments

type
character(1): Either "shows" or "movies". For season/episode-specific functions, values seasons or episodes are also allowed.

limit
integer(1) [10L]: Number of items to return. Must be greater than 0 and will be coerced to integer.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

query
character(1): Search titles and descriptions.

years
integer(1): 4-digit year or range, e.g. 2010-2020. Can also be an integer vector of length two which will be coerced appropriately.

genres
character(n): Genre slug(s). See genres for a table of genres. Multiple values are allowed and will be concatenated.

languages
character(n): Two-digit language code(s). Also see languages for table of available languages.

countries
character(n): Two-character country code(s). See countries.

runtimes
character | integer: Integer range in minutes, e.g. 30-90. Can also be an integer vector of length two which will be coerced appropriately.

ratings
character | integer: Integer range between 0 and 100. Can also be an integer vector of length two which will be coerced appropriately.

certifications
character(n): Certification(s) like pg-13. Multiple values are allowed. Use certifications for reference. Note that there are different certifications for shows and movies.

networks
character(n): (Shows only) Network name like HBO. See networks for a list of known networks.

status
character(n): (Shows only) The status of the shows. One of returning series, in production, planned, canceled, or ended

period
character(1) ["weekly"]: Which period to filter by. Possible values are "weekly", "monthly", "yearly", "all".

start_date
character(1): A date in the past from which on to count updates. If no date is supplied, the default is to use the date 7 days in the past relative to the current date. Value must either be standard YYYY-MM-DD format or an object of class Date, which will then be coerced via as.character().

Value

A tibble().

Details

These functions access the automatically updated lists provided by trakt.tv. Each function comes in two flavors: Shows or movies. The following descriptions are adapted directly from the API reference.

Popular: Popularity is calculated using the rating percentage and the number of ratings.

Trending: Returns all movies/shows being watched right now. Movies/shows with the most users are returned first.

Played: Returns the most played (a single user can watch multiple times) movies/shows in the specified time period.

Watched: Returns the most watched (unique users) movies/shows in the specified time period.

Collected: Returns the most collected (unique users) movies/shows in the specified time period.

Anticipated: Returns the most anticipated movies/shows based on the number of lists a movie/show appears on.

Updates: Returns all movies/shows updated since the specified UTC start_date. In this case, the upper bound for limit is 100.

The functions for Played, Watched, Collected and Played each return the same additional variables besides the media information: watcher_count, play_count, collected_count, collector_count.

Note

trakt.updates does not support filters.

See also

Examples

# \dontrun{
# Get popular shows with only ids
trakt.popular(type = "shows")

#> # A tibble: 10 x 7
#>    title                  year trakt  slug                  tvdb  imdb     tmdb 
#>    <chr>                 <int> <chr>  <chr>                 <chr> <chr>    <chr>
#>  1 Game of Thrones        2011 1390   game-of-thrones       1213… tt09449… 1399 
#>  2 Breaking Bad           2008 1388   breaking-bad          81189 tt09037… 1396 
#>  3 The Walking Dead       2010 1393   the-walking-dead      1530… tt15202… 1402 
#>  4 The Big Bang Theory    2007 1409   the-big-bang-theory   80379 tt08982… 1418 
#>  5 Sherlock               2010 19792  sherlock              1769… tt14755… 19885
#>  6 How I Met Your Mother  2005 1095   how-i-met-your-mother 75760 tt04606… 1100 
#>  7 Dexter                 2006 1396   dexter                79349 tt07732… 1405 
#>  8 Arrow                  2012 1403   arrow                 2576… tt21930… 1412 
#>  9 Friends                1994 1657   friends               79168 tt01087… 1668 
#> 10 Stranger Things        2016 104439 stranger-things       3052… tt45743… 66732

# Get trending movies with extended information
trakt.trending(type = "movies", 5, extended = "full")

#> # A tibble: 5 x 22
#>   watchers title  year tagline overview released   runtime country trailer
#>      <int> <chr> <int> <chr>   <chr>    <date>       <int> <chr>   <chr>  
#> 1      151 Spid…  2019 It’s t… Peter P… 2019-07-02     129 us      http:/…
#> 2      151 Fast…  2019 ""      A spino… 2019-08-02     136 us      http:/…
#> 3       84 Toy …  2019 Get Re… "Woody … 2019-06-21     100 us      http:/…
#> 4       67 John…  2019 The Gr… Super-a… 2019-05-17     131 us      http:/…
#> 5       61 Mids…  2019 Let th… A young… 2019-07-03     147 us      http:/…
#> # … with 13 more variables: homepage <chr>, rating <dbl>, votes <int>,
#> #   comment_count <int>, updated_at <dttm>, language <chr>,
#> #   available_translations <list>, genres <list>, certification <chr>,
#> #   trakt <int>, slug <chr>, imdb <chr>, tmdb <int>

# Get top 5 anticipated movies
trakt.anticipated(type = "movies", 5)

#> # A tibble: 5 x 7
#>   list_count title                year  trakt slug                 imdb     tmdb
#>        <int> <chr>               <int>  <int> <chr>                <chr>   <int>
#> 1      31352 Avatar 2             2021  56580 avatar-2-2021        tt163…  76600
#> 2      31215 Star Wars: The Ris…  2019 114335 star-wars-the-rise-… tt252… 181812
#> 3      23270 The Flash            2021 195144 the-flash-2021       tt043… 298618
#> 4      15623 Guardians of the G…  2022 293990 guardians-of-the-ga… tt679… 447365
#> 5      15541 Sherlock Holmes 3    2021 191791 sherlock-holmes-3-2… tt209… 297708

# Get 15 the most anticipated upcoming shows on Netflix that air this year
current_year <- format(Sys.Date(), "%Y")
trakt.anticipated("shows", limit = 15, networks = "Netflix", years = current_year)

#> # A tibble: 15 x 8
#>    list_count title              year trakt  slug            tvdb  imdb    tmdb 
#>         <int> <chr>             <int> <chr>  <chr>           <chr> <chr>   <chr>
#>  1       9072 The Witcher        2019 138163 the-witcher-20… 3626… tt5180… 71912
#>  2        997 Raising Dion       2019 124235 raising-dion    3359… tt7826… 93392
#>  3        869 Living With Your…  2019 151776 living-with-yo… 3685… tt8880… 81532
#>  4        427 Daybreak           2019 152108 daybreak-2019   3547… tt8755… 83478
#>  5        140 V-Wars             2019 133365 v-wars          3493… tt7403… <NA> 
#>  6        139 Sintonia           2019 131399 sintonia        3443… tt8165… <NA> 
#>  7        116 The Faction        2019 141592 the-faction     3566… <NA>    <NA> 
#>  8         49 Rhythm + Flow      2019 152211 rhythm-flow     3690… tt9278… 83996
#>  9         29 Hache              2019 147926 hache           3555… tt8575… 84112
#> 10         25 Seis Manos         2019 152707 seis-manos      3665… tt1010… 93953
#> 11         19 Cidades Invisive…  2019 134761 cidades-invisi… 3493… <NA>    <NA> 
#> 12          9 Living Undocumen…  2019 153094 living-undocum… 3699… tt1097… <NA> 
#> 13          7 Brotherhood        2019 153095 brotherhood-20… 3589… tt8905… <NA> 
#> 14          7 Queer Eye: We're…  2019 152948 queer-eye-we-r… 3697… <NA>    <NA> 
#> 15          7 Prank Encounters   2019 152904 prank-encounte… 3697… <NA>    <NA> 

# Get the most popular German-language movies between 1990 and 2010
trakt.popular("movies", languages = "de", years = c(1990, 2010))

#> # A tibble: 10 x 6
#>    title                    year trakt slug                       imdb      tmdb
#>    <chr>                   <int> <int> <chr>                      <chr>    <int>
#>  1 Downfall                 2004   489 downfall-2004              tt03631…   613
#>  2 The Lives of Others      2006   461 the-lives-of-others-2006   tt04050…   582
#>  3 Run Lola Run             1998    73 run-lola-run-1998          tt01308…   104
#>  4 The Wave                 2008  3954 the-wave-2008              tt10636…  7735
#>  5 Good bye, Lenin!         2003   285 good-bye-lenin-2003        tt03013…   338
#>  6 The Experiment           2001   454 the-experiment-2001        tt02502…   575
#>  7 Funny Games              1997  5445 funny-games-1997           tt01191… 10234
#>  8 The White Ribbon         2009 24371 the-white-ribbon-2009      tt11493… 37903
#>  9 The Baader Meinhof Com…  2008  3789 the-baader-meinhof-comple… tt07654…  6968
#> 10 Cargo                    2009 21807 cargo-2009                 tt03819… 34069

# }