Get a single comment

Get a single comment

comments_comment(id, extended = c("min", "full"))

comments_replies(id, extended = c("min", "full"))

comments_likes(id, extended = c("min", "full"))

comments_item(id, extended = c("min", "full"))

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things", package = "tRakt") for more details.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things", package = "tRakt") for more details.

Source

comments_comment() wraps endpoint comments/:id.

comments_replies() wraps endpoint comments/:id/replies.

comments_likes() wraps endpoint comments/:id/likes.

comments_item() wraps endpoint comments/:id/item.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

Functions

comments_replies: Get a comment's replies

comments_likes: Get users who liked a comment.

comments_item: Get the media item attached to the comment.

See also

Examples

# A single comment
comments_comment("236397")

#> # A tibble: 1 x 16
#>       id comment spoiler review parent_id created_at         
#>    <int> <chr>   <lgl>   <lgl>      <int> <dttm>             
#> 1 236397 All th… FALSE   FALSE          0 2019-06-09 21:33:00
#> # … with 10 more variables: updated_at <dttm>, replies <int>, likes <int>,
#> #   user_rating <int>, username <chr>, private <lgl>, user_name <chr>,
#> #   vip <lgl>, vip_ep <lgl>, user_slug <chr>

# Multiple comments
comments_comment(c("236397", "112561"))

#> # A tibble: 2 x 16
#>       id comment spoiler review parent_id created_at         
#>    <int> <chr>   <lgl>   <lgl>      <int> <dttm>             
#> 1 236397 All th… FALSE   FALSE          0 2019-06-09 21:33:00
#> 2 112561 Seriou… FALSE   FALSE          0 2017-01-31 17:48:59
#> # … with 10 more variables: updated_at <dttm>, replies <int>, likes <int>,
#> #   user_rating <int>, username <chr>, private <lgl>, user_name <chr>,
#> #   vip <lgl>, vip_ep <lgl>, user_slug <chr>

# \dontrun{
comments_replies("236397")

#> # A tibble: 1 x 16
#>       id comment spoiler review parent_id created_at         
#>    <int> <chr>   <lgl>   <lgl>      <int> <dttm>             
#> 1 250976 "@jemu… FALSE   FALSE     236397 2019-09-18 02:57:44
#> # … with 10 more variables: updated_at <dttm>, replies <int>, likes <int>,
#> #   user_rating <int>, username <chr>, private <lgl>, user_name <chr>,
#> #   vip <lgl>, vip_ep <lgl>, user_slug <chr>

# }
# \dontrun{
comments_likes("236397")

#> # A tibble: 2 x 7
#>   liked_at                 username   private user_name  vip   vip_ep user_slug 
#>   <chr>                    <chr>      <lgl>   <chr>      <lgl> <lgl>  <chr>     
#> 1 2019-06-29T21:57:36.000Z Khawlah    FALSE   Khawlah    FALSE FALSE  khawlah   
#> 2 2019-09-15T02:43:31.000Z Outlanders FALSE   Outlanders FALSE FALSE  outlanders

# }
# \dontrun{
# A movie
comments_item("236397")

#> # A tibble: 1 x 7
#>   type  title                   year trakt  slug                   imdb    tmdb 
#>   <chr> <chr>                  <int> <chr>  <chr>                  <chr>   <chr>
#> 1 movie John Wick: Chapter 3 …  2019 304278 john-wick-chapter-3-p… tt6146… 4581…

comments_item("236397", extended = "full")

#> # A tibble: 1 x 23
#>   type  title  year tagline overview released   runtime country trailer homepage
#>   <chr> <chr> <int> <chr>   <chr>    <date>       <int> <chr>   <chr>   <chr>   
#> 1 movie John…  2019 The Gr… Super-a… 2019-05-17     131 us      http:/… https:/…
#> # … with 13 more variables: status <chr>, rating <dbl>, votes <int>,
#> #   comment_count <int>, updated_at <dttm>, language <chr>,
#> #   available_translations <list>, genres <list>, certification <chr>,
#> #   trakt <chr>, slug <chr>, imdb <chr>, tmdb <chr>

# A show
comments_item("120768")

#> # A tibble: 1 x 8
#>   type  title           year trakt  slug           tvdb   imdb      tmdb 
#>   <chr> <chr>          <int> <chr>  <chr>          <chr>  <chr>     <chr>
#> 1 show  13 Reasons Why  2017 116129 13-reasons-why 323168 tt1837492 66788

comments_item("120768", extended = "full")

#> # A tibble: 1 x 28
#>   type  title  year overview first_aired         runtime certification country
#>   <chr> <chr> <int> <chr>    <dttm>                <int> <chr>         <chr>  
#> 1 show  13 R…  2017 Follows… 2017-03-31 16:00:00      57 TV-MA         us     
#> # … with 20 more variables: trailer <chr>, homepage <chr>, status <chr>,
#> #   rating <dbl>, votes <int>, comment_count <int>, network <chr>,
#> #   updated_at <dttm>, language <chr>, available_translations <list>,
#> #   genres <list>, aired_episodes <int>, airs_day <chr>, airs_time <chr>,
#> #   airs_timezone <chr>, trakt <chr>, slug <chr>, tvdb <chr>, imdb <chr>,
#> #   tmdb <chr>

# A season
comments_item("140265")

#> # A tibble: 1 x 12
#>   type  title  year trakt slug  tvdb  imdb  tmdb  season season_trakt
#>   <chr> <chr> <int> <chr> <chr> <chr> <chr> <chr>  <int> <chr>       
#> 1 seas… Twin…  1990 1907  twin… 70533 tt00… 1920       3 138350      
#> # … with 2 more variables: season_tvdb <chr>, season_tmdb <chr>

comments_item("140265", extended = "full")

#> # A tibble: 1 x 40
#>   type  title  year overview first_aired         runtime certification country
#>   <chr> <chr> <int> <chr>    <dttm>                <int> <chr>         <chr>  
#> 1 seas… Twin…  1990 The bod… 1990-04-09 00:00:00      46 TV-MA         us     
#> # … with 32 more variables: trailer <chr>, homepage <chr>, status <chr>,
#> #   rating <dbl>, votes <int>, comment_count <int>, network <chr>,
#> #   updated_at <dttm>, language <chr>, available_translations <list>,
#> #   genres <list>, aired_episodes <int>, airs_day <chr>, airs_time <chr>,
#> #   airs_timezone <chr>, trakt <chr>, slug <chr>, tvdb <chr>, imdb <chr>,
#> #   tmdb <chr>, season <int>, season_rating <dbl>, season_votes <int>,
#> #   season_episode_count <int>, season_aired_episodes <int>,
#> #   season_title <chr>, season_overview <chr>, season_first_aired <chr>,
#> #   season_network <chr>, season_trakt <chr>, season_tvdb <chr>,
#> #   season_tmdb <chr>

# An episode
comments_item("136632")

#> # A tibble: 1 x 15
#>   type  title  year trakt slug  tvdb  imdb  tmdb  season episode episode_title
#>   <chr> <chr> <int> <chr> <chr> <chr> <chr> <chr>  <int>   <int> <chr>        
#> 1 epis… Game…  2011 1390  game… 1213… tt09… 1399       7       4 The Spoils o…
#> # … with 4 more variables: episode_trakt <chr>, episode_tvdb <chr>,
#> #   episode_imdb <chr>, episode_tmdb <chr>

comments_item("136632", extended = "full")

#> # A tibble: 1 x 44
#>   type  title  year overview first_aired         runtime certification country
#>   <chr> <chr> <int> <chr>    <dttm>                <int> <chr>         <chr>  
#> 1 epis… Game…  2011 Seven n… 2011-04-18 01:00:00      60 TV-MA         us     
#> # … with 36 more variables: trailer <chr>, homepage <chr>, status <chr>,
#> #   rating <dbl>, votes <int>, comment_count <int>, network <chr>,
#> #   updated_at <dttm>, language <chr>, available_translations <list>,
#> #   genres <list>, aired_episodes <int>, airs_day <chr>, airs_time <chr>,
#> #   airs_timezone <chr>, trakt <chr>, slug <chr>, tvdb <chr>, imdb <chr>,
#> #   tmdb <chr>, season <int>, episode <int>, episode_title <chr>,
#> #   episode_number_abs <int>, episode_overview <chr>, episode_rating <dbl>,
#> #   episode_votes <int>, episode_comment_count <int>,
#> #   episode_first_aired <chr>, episode_updated_at <chr>,
#> #   episode_available_translations <list>, episode_runtime <int>,
#> #   episode_trakt <chr>, episode_tvdb <chr>, episode_imdb <chr>,
#> #   episode_tmdb <chr>

# }