Get all movie / show aliases

Get all movie / show aliases

Usage

movies_aliases(id)

shows_aliases(id)

Source

movies_aliases() wraps endpoint /movies/:id/aliases.

shows_aliases() wraps endpoint /shows/:id/aliases.

Arguments

id

character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("tRakt") for more details.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

See also

Examples

movies_aliases(190430)
#> # A tibble: 30 × 2
#>    title                   country
#>    <chr>                   <chr>  
#>  1 X-Men Origins: Deadpool us     
#>  2 X-Men: Deadpool         us     
#>  3 Deadpool                br     
#>  4 Дедпул                  bg     
#>  5 Dedpul                  rs     
#>  6 Дэдпул                  ru     
#>  7 惡棍英雄：死侍          tw     
#>  8 死侍：不死现身          cn     
#>  9 X战警：死侍             cn     
#> 10 Deadpool 1              us     
#> # ℹ 20 more rows
shows_aliases(104439)
#> # A tibble: 711 × 2
#>    title                   country
#>    <chr>                   <chr>  
#>  1 Очень странные дела     us     
#>  2 Странные вещи           us     
#>  3 Загадкові справи        ua     
#>  4 Keisti dalykai          lt     
#>  5 Крайне странные события ru     
#>  6 უცნაური საქმეები        ge     
#>  7 Stranger Things (2016)  cn     
#>  8 Stranger Things         cn     
#>  9 Странные вещи           ru     
#> 10 Крайне странные события ru     
#> # ℹ 701 more rows