Get all comments of a thing

Get all comments of a thing

movies_comments(id, sort = c("newest", "oldest", "likes", "replies"),
  extended = c("min", "full"), limit = 10L)

shows_comments(id, sort = c("newest", "oldest", "likes", "replies"),
  extended = c("min", "full"), limit = 10L)

seasons_comments(id, season = 1L, sort = c("newest", "oldest", "likes",
  "replies"), extended = c("min", "full"), limit = 10L)

episodes_comments(id, season = 1L, episode = 1L, sort = c("newest",
  "oldest", "likes", "replies"), extended = c("min", "full"),
  limit = 10L)

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things", package = "tRakt") for more details.

sort
character(1) ["newest"]: Comment sort order, one of "newest", "oldest", "likes" or "replies".

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things", package = "tRakt") for more details.

limit
integer(1) [10L]: Number of items to return. Must be greater than 0 and will be coerced via as.integer().

season
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

episode
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Source

movies/:id/comments/:sort

shows/:id/comments/:sort

shows/:id/seasons/:season/comments/:sort

shows/:id/seasons/:season/episodes/:episode/comments/:sort

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

Functions

movies_comments: Get comments for a movie

shows_comments: Get comments for a movie

seasons_comments: Get comments for a season

episodes_comments: Get comments for an episode

See also

Examples

# \dontrun{
movies_comments(193972)

#> # A tibble: 10 x 16
#>        id comment spoiler review parent_id created_at         
#>     <int> <chr>   <lgl>   <lgl>      <int> <dttm>             
#>  1 254777 I was … FALSE   FALSE          0 2019-10-12 16:41:10
#>  2 254536 It's a… TRUE    FALSE          0 2019-10-11 15:20:59
#>  3 252661 how do… FALSE   FALSE          0 2019-09-30 00:56:09
#>  4 252389 "One t… FALSE   FALSE          0 2019-09-28 18:33:46
#>  5 252274 Woody … FALSE   FALSE          0 2019-09-28 00:37:38
#>  6 252267 Can't … FALSE   FALSE          0 2019-09-27 22:32:29
#>  7 251391 "that … FALSE   FALSE          0 2019-09-21 16:42:50
#>  8 250960 i grew… FALSE   FALSE          0 2019-09-18 00:26:55
#>  9 250798 It's v… FALSE   FALSE          0 2019-09-16 16:48:08
#> 10 250646 "It's … FALSE   FALSE          0 2019-09-15 16:31:24
#> # … with 10 more variables: updated_at <dttm>, replies <int>, likes <int>,
#> #   user_rating <int>, username <chr>, private <lgl>, user_name <chr>,
#> #   vip <lgl>, vip_ep <lgl>, user_slug <chr>

shows_comments(46241, sort = "likes")

#> # A tibble: 10 x 16
#>        id comment spoiler review parent_id created_at         
#>     <int> <chr>   <lgl>   <lgl>      <int> <dttm>             
#>  1   4528 The br… FALSE   FALSE          0 2014-12-27 21:04:43
#>  2  11599 Televi… FALSE   FALSE          0 2013-02-24 18:13:58
#>  3  39755 Holy h… FALSE   FALSE          0 2015-03-11 07:54:07
#>  4 143277 "After… FALSE   FALSE          0 2017-10-01 18:42:14
#>  5  34510 It's O… FALSE   FALSE          0 2014-07-31 22:21:49
#>  6  42940 "Looke… FALSE   FALSE          0 2015-05-06 17:25:07
#>  7  11140 I like… FALSE   FALSE          0 2014-12-13 17:56:40
#>  8 185427 "In Fe… FALSE   FALSE          0 2018-08-01 12:50:58
#>  9  71014 Watch … FALSE   FALSE          0 2016-02-10 04:16:54
#> 10  38229 "This … FALSE   FALSE          0 2015-02-09 07:47:51
#> # … with 10 more variables: updated_at <dttm>, replies <int>, likes <int>,
#> #   user_rating <int>, username <chr>, private <lgl>, user_name <chr>,
#> #   vip <lgl>, vip_ep <lgl>, user_slug <chr>

seasons_comments(46241, season = 1, sort = "likes")

#> # A tibble: 0 x 0

episodes_comments(46241, season = 1, episode = 2, sort = "likes")

#> # A tibble: 0 x 0

# }