Get lists containing a movie, show, season, episode or person

Get lists containing a movie, show, season, episode or person

Usage

movies_lists(
  id,
  type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L,
  extended = c("min", "full")
)

shows_lists(
  id,
  type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L,
  extended = c("min", "full")
)

seasons_lists(
  id,
  season,
  type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L,
  extended = c("min", "full")
)

episodes_lists(
  id,
  season,
  episode,
  type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L,
  extended = c("min", "full")
)

people_lists(
  id,
  type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L,
  extended = c("min", "full")
)

Source

movies_lists() wraps endpoint /movies/:id/lists/:type/:sort.

shows_lists() wraps endpoint /shows/:id/lists/:type/:sort.

seasons_lists() wraps endpoint /shows/:id/seasons/:season/lists/:type/:sort.

episodes_lists() wraps endpoint /shows/:id/seasons/:season/episodes/:episode/lists/:type/:sort.

Arguments

id

character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("tRakt") for more details.

type

character(1) ["all"]: The type of list, one of "all", "personal", "official" or "watchlists".

sort

character(1) ["popular"]: Sort lists by one of "popular", "likes", "comments", "items", "added" or "updated".

limit

integer(1) [10L]: Number of items to return. Must be greater than 0 and will be coerced via as.integer().

extended

character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("tRakt") for more details.

season, episode

integer(1) [1L]: The season and episode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

Functions

movies_lists(): Lists containing a movie.

shows_lists(): Lists containing a show.

seasons_lists(): Lists containing a season.

episodes_lists(): Lists containing an episode.

people_lists(): Lists containing a person.

See also

Examples

if (FALSE) { # \dontrun{
movies_lists("190430", type = "personal", limit = 5)
} # }
if (FALSE) { # \dontrun{
shows_lists("46241")
} # }
if (FALSE) { # \dontrun{
seasons_lists("46241", season = 1)
} # }
if (FALSE) { # \dontrun{
episodes_lists("46241", season = 1, episode = 1)
} # }
if (FALSE) { # \dontrun{
people_lists("david-tennant")

people_lists("emilia-clarke", sort = "items")
} # }