Get lists containing a movie, show, season, episode or person

Get lists containing a movie, show, season, episode or person

movies_lists(id, type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L, extended = c("min", "full"))

shows_lists(id, type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L, extended = c("min", "full"))

seasons_lists(id, season, type = c("all", "personal", "official",
  "watchlists"), sort = c("popular", "likes", "comments", "items",
  "added", "updated"), limit = 10L, extended = c("min", "full"))

episodes_lists(id, season, episode, type = c("all", "personal",
  "official", "watchlists"), sort = c("popular", "likes", "comments",
  "items", "added", "updated"), limit = 10L, extended = c("min",
  "full"))

people_lists(id, type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L, extended = c("min", "full"))

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things", package = "tRakt") for more details.

type
character(1) ["all"]: The type of list, one of "all", "personal", "official" or "watchlists".

sort
character(1) ["popular"]: Sort lists by one of "popular", "likes", "comments", "items", "added" or "updated".

limit
integer(1) [10L]: Number of items to return. Must be greater than 0 and will be coerced via as.integer().

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things", package = "tRakt") for more details.

season
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

episode
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

Functions

movies_lists: Lists containing a movie.

shows_lists: Lists containing a show.

seasons_lists: Lists containing a season.

episodes_lists: Lists containing an episode.

people_lists: Lists containing a person.

See also

Examples

# \dontrun{
movies_lists("190430", type = "personal", limit = 5)

#> # A tibble: 5 x 20
#>   name  description privacy display_numbers allow_comments sort_by sort_how
#>   <chr> <chr>       <chr>   <lgl>           <lgl>          <chr>   <chr>   
#> 1 Imdb… Daily Refr… public  TRUE            TRUE           rank    asc     
#> 2 Come… ""          public  FALSE           FALSE          rank    asc     
#> 3 Acti… ""          public  FALSE           TRUE           releas… desc    
#> 4 Movi… ""          public  FALSE           TRUE           added   desc    
#> 5 Acti… ""          public  FALSE           TRUE           title   asc     
#> # … with 13 more variables: created_at <dttm>, updated_at <dttm>,
#> #   item_count <int>, comment_count <int>, likes <int>, trakt <chr>,
#> #   slug <chr>, username <chr>, private <lgl>, user_name <chr>, vip <lgl>,
#> #   vip_ep <lgl>, user_slug <chr>

shows_lists("46241")

#> # A tibble: 10 x 20
#>    name  description privacy display_numbers allow_comments sort_by sort_how
#>    <chr> <chr>       <chr>   <lgl>           <lgl>          <chr>   <chr>   
#>  1 Trak… The Trakt … public  TRUE            TRUE           rank    asc     
#>  2 Imdb… Daily Refr… public  TRUE            TRUE           rank    asc     
#>  3 Scie… ""          public  FALSE           FALSE          releas… desc    
#>  4 Chan… Shows on C… public  TRUE            TRUE           rank    asc     
#>  5 All … List creat… public  FALSE           FALSE          rank    asc     
#>  6 Brit… TV series … public  FALSE           TRUE           title   asc     
#>  7 All … ""          public  FALSE           TRUE           rank    asc     
#>  8 TV    TV Shows    public  FALSE           TRUE           title   asc     
#>  9 WATC… ""          public  TRUE            FALSE          rank    asc     
#> 10 TV S… ""          public  FALSE           TRUE           title   asc     
#> # … with 13 more variables: created_at <dttm>, updated_at <dttm>,
#> #   item_count <int>, comment_count <int>, likes <int>, trakt <chr>,
#> #   slug <chr>, username <chr>, private <lgl>, user_name <chr>, vip <lgl>,
#> #   vip_ep <lgl>, user_slug <chr>

seasons_lists("46241", season = 1)

#> # A tibble: 10 x 20
#>    name  description privacy display_numbers allow_comments sort_by sort_how
#>    <chr> <chr>       <chr>   <lgl>           <lgl>          <chr>   <chr>   
#>  1 OMG!… "My list o… public  FALSE           TRUE           title   desc    
#>  2 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#>  3 show… ""          public  FALSE           TRUE           runtime desc    
#>  4 Down… tv shows    public  FALSE           TRUE           title   asc     
#>  5 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#>  6 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#>  7 My O… "==  /!\\ … public  TRUE            TRUE           rank    asc     
#>  8 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#>  9 tv s… ""          public  FALSE           TRUE           rank    asc     
#> 10 Show… ""          public  FALSE           TRUE           rank    asc     
#> # … with 13 more variables: created_at <dttm>, updated_at <dttm>,
#> #   item_count <int>, comment_count <int>, likes <int>, trakt <chr>,
#> #   slug <chr>, username <chr>, private <lgl>, user_name <chr>, vip <lgl>,
#> #   vip_ep <lgl>, user_slug <chr>

episodes_lists("46241", season = 1, episode = 1)

#> # A tibble: 10 x 20
#>    name  description privacy display_numbers allow_comments sort_by sort_how
#>    <chr> <chr>       <chr>   <lgl>           <lgl>          <chr>   <chr>   
#>  1 seri… tg          public  FALSE           TRUE           rank    asc     
#>  2 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#>  3 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#>  4 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#>  5 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#>  6 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#>  7 TV 2… "All episo… public  FALSE           TRUE           added   desc    
#>  8 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#>  9 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#> 10 Watc… Movies, sh… <NA>    FALSE           FALSE          rank    asc     
#> # … with 13 more variables: created_at <dttm>, updated_at <dttm>,
#> #   item_count <int>, comment_count <int>, likes <int>, trakt <chr>,
#> #   slug <chr>, username <chr>, private <lgl>, user_name <chr>, vip <lgl>,
#> #   vip_ep <lgl>, user_slug <chr>

people_lists("david-tennant")

#> # A tibble: 10 x 20
#>    name  description privacy display_numbers allow_comments sort_by sort_how
#>    <chr> <chr>       <chr>   <lgl>           <lgl>          <chr>   <chr>   
#>  1 Acto… ""          public  FALSE           FALSE          rank    asc     
#>  2 Moni… ""          public  FALSE           FALSE          rank    asc     
#>  3 Acto… ""          public  FALSE           TRUE           rank    asc     
#>  4 Faves ""          public  FALSE           FALSE          rank    asc     
#>  5 favo… ""          public  FALSE           FALSE          added   desc    
#>  6 Best… ""          public  FALSE           FALSE          rank    asc     
#>  7 Neve… ""          public  FALSE           TRUE           random  asc     
#>  8 Peop… People who… public  FALSE           FALSE          releas… desc    
#>  9 Favo… Acting: I … public  FALSE           TRUE           random  asc     
#> 10 Acto… ""          public  TRUE            TRUE           rank    asc     
#> # … with 13 more variables: created_at <dttm>, updated_at <dttm>,
#> #   item_count <int>, comment_count <int>, likes <int>, trakt <chr>,
#> #   slug <chr>, username <chr>, private <lgl>, user_name <chr>, vip <lgl>,
#> #   vip_ep <lgl>, user_slug <chr>

# }