Get the cast and crew of a show or movie

Returns all cast and crew for a show/movie, depending on how much data is available.

trakt.shows.people(target, extended = c("min", "full"))

trakt.movies.people(target, extended = c("min", "full"))

Arguments

target
character(1): The ID of the item requested. Preferably the Trakt ID (e.g. 1390), slug (e.g. "game-of-thrones"), or IMDb ID (e.g. "tt0944947"). Can also be of length greater than 1, in which case the function is called on all target values separately and the result is combined.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

Value

A list of one or more tibbles for cast and crew. The latter tibble objects are as flat as possible.

See also

Examples

trakt.shows.people("breaking-bad")

#> $cast
#> # A tibble: 8 x 6
#>   character      name        trakt  slug                          imdb    tmdb  
#>   <chr>          <chr>       <chr>  <chr>                         <chr>   <chr> 
#> 1 Walter White   Bryan Cran… 297737 bryan-cranston                nm0186… 17419 
#> 2 Skyler White   Anna Gunn   420986 anna-gunn                     nm0348… 134531
#> 3 Jesse Pinkman  Aaron Paul  415249 aaron-paul                    nm0666… 84497 
#> 4 Hank Schrader  Dean Norris 415192 dean-norris                   nm0606… 14329 
#> 5 Marie Schrader Betsy Bran… 436547 betsy-brandt                  nm1336… 12179…
#> 6 Walter White … RJ Mitte    459150 rj-mitte-06c377f6-88c1-4969-… nm2666… 209674
#> 7 Mike Ehrmantr… Jonathan B… 411251 jonathan-banks                nm0052… 783   
#> 8 Saul Goodman   Bob Odenki… 294857 bob-odenkirk                  nm0644… 59410 
#> 
#> $crew
#> # A tibble: 18 x 7
#>    job          name        trakt  slug                 imdb   tmdb  crew_type  
#>    <chr>        <chr>       <chr>  <chr>                <chr>  <chr> <chr>      
#>  1 Executive P… Michelle M… 12427  michelle-maclaren    nm053… 29779 production 
#>  2 Producer     Peter Gould 20811  peter-gould          nm033… 24951 production 
#>  3 Producer     Thomas Sch… 67528  thomas-schnauz       nm104… 1030… production 
#>  4 Producer     Bryan Cran… 297737 bryan-cranston       nm018… 17419 production 
#>  5 Producer     Stewart Ly… 325805 stewart-lyons        nm052… 1223… production 
#>  6 Producer     Moira Wall… 325808 moira-walley-beckett nm090… 1223… production 
#>  7 Producer     Karen Moore 365840 karen-moore          nm102… 1223… production 
#>  8 Producer     Patty Lin   398571 patty-lin            nm051… 1218… production 
#>  9 Producer     Melissa Be… 400288 melissa-bernstein    nm221… 1223… production 
#> 10 Producer     George Mas… 400289 george-mastras       nm229… 1223… production 
#> 11 Producer     Diane Merc… 400290 diane-mercer         nm103… 1223… production 
#> 12 Executive P… Mark Johns… 412662 mark-johnson-15eae8… nm042… 5162  production 
#> 13 Producer     John Shiban 418282 john-shiban          nm079… 29924 production 
#> 14 Executive P… Vince Gill… 423003 vince-gilligan       nm031… 66633 production 
#> 15 Producer     Sam Catlin  436598 sam-catlin           nm102… 1223… production 
#> 16 Production … Robb Wilso… 15105  robb-wilson-king     nm045… 21640 art        
#> 17 Original Mu… Dave Porter 241567 dave-porter          nm172… 1280… sound      
#> 18 Costume Des… Kathleen D… 27184  kathleen-detoro      nm022… 35583 costume & …
#>