Get the cast and crew of a show or movie

Returns all cast and crew for a show/movie, depending on how much data is available.

movies_people(id, extended = c("min", "full"))

shows_people(id, guest_stars = FALSE, extended = c("min", "full"))

seasons_people(
  id,
  season = 1L,
  guest_stars = FALSE,
  extended = c("min", "full")
)

episodes_people(
  id,
  season = 1L,
  episode = 1L,
  guest_stars = FALSE,
  extended = c("min", "full")
)

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things", package = "tRakt") for more details.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things", package = "tRakt") for more details.

guest_stars
logical(1) ["FALSE"]: Also include guest stars. This returns a lot of data, so use with care.

season
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

episode
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Source

movies_people() wraps endpoint movies/:id/people.

shows_people() wraps endpoint shows/:id/people.

seasons_people() wraps endpoint shows/:id/seasons/:season/people.

episodes_people() wraps endpoint shows/:id/seasons/:season/episodes/:episode/people.

Value

A list of one or more tibbles for cast and/or crew. The latter tibble objects are as flat as possible.

Note

As of 2019-09-30, there are two representations of character[s] and job[s]: One is a regular character variable, and the other is a list-column. The former is deprecated and only included for compatibility reasons.

See also

Examples

# \dontrun{
movies_people("deadpool-2016")

#> $cast
#> # A tibble: 47 x 7
#>    character               characters name        trakt  slug       imdb   tmdb 
#>    <chr>                   <list>     <chr>       <chr>  <chr>      <chr>  <chr>
#>  1 Wade Wilson, Deadpool   <chr [2]>  Ryan Reyno… 416707 ryan-reyn… nm000… 10859
#>  2 Dopinder                <chr [1]>  Karan Soni  203567 karan-soni nm417… 1109…
#>  3 Francis Freeman, Ajax   <chr [2]>  Ed Skrein   456186 ed-skrein  nm453… 1047…
#>  4 Warlord                 <chr [1]>  Michael Be… 417888 michael-b… nm007… 27588
#>  5 Piotr Rasputin, Coloss… <chr [2]>  Stefan Kap… 431218 stefan-ka… nm129… 80507
#>  6 Ellie Phimister, Negas… <chr [2]>  Brianna Hi… 554869 brianna-h… nm655… 1492…
#>  7 Jeremy (Pizza Guy)      <chr [1]>  Style Dayne 495632 style-day… nm440… 1417…
#>  8 Gavin Merchant          <chr [1]>  Kyle Cassie 47532  kyle-cass… nm100… 62915
#>  9 Meghan Orlovsky         <chr [1]>  Taylor Hic… 554871 taylor-hi… nm688… 1493…
#> 10 Teen Girl #1            <chr [1]>  Ayzee       710498 ayzee      <NA>   1683…
#> # … with 37 more rows
#> 
#> $crew
#> # A tibble: 83 x 8
#>    job          jobs    name          trakt  slug         imdb   tmdb  crew_type
#>    <chr>        <list>  <chr>         <chr>  <chr>        <chr>  <chr> <chr>    
#>  1 Casting      <chr [… Ronna Kress   638    ronna-kress  nm000… 1720  producti…
#>  2 Producer     <chr [… Lauren Shule… 5528   lauren-shul… nm079… 7200  producti…
#>  3 Executive P… <chr [… Stan Lee      16093  stan-lee     nm049… 7624  producti…
#>  4 Executive P… <chr [… Paul Wernick  78174  paul-wernick nm111… 91269 producti…
#>  5 Casting      <chr [… Corinne Clark 179298 corinne-cla… nm016… 9862… producti…
#>  6 Casting      <chr [… Jennifer Page 179299 jennifer-pa… nm065… 1122… producti…
#>  7 Producer     <chr [… Simon Kinberg 414520 simon-kinbe… nm133… 11092 producti…
#>  8 Producer     <chr [… Ryan Reynolds 416707 ryan-reynol… nm000… 10859 producti…
#>  9 Casting Ass… <chr [… Hannah d'Ang… 548376 hannah-d-an… nm378… 1466… producti…
#> 10 Production … <chr [… Sean Haworth  6957   sean-haworth nm037… 9618  art      
#> # … with 73 more rows
#> 

shows_people("breaking-bad")

#> $cast
#> # A tibble: 9 x 8
#>   character    characters episode_count name     trakt  slug         imdb  tmdb 
#>   <chr>        <list>             <int> <chr>    <chr>  <chr>        <chr> <chr>
#> 1 "Walter Whi… <chr [1]>             62 Bryan C… 297737 bryan-crans… nm01… 17419
#> 2 "Jesse Pink… <chr [1]>             62 Aaron P… 415249 aaron-paul   nm06… 84497
#> 3 "Skyler Whi… <chr [1]>             62 Anna Gu… 420986 anna-gunn    nm03… 1345…
#> 4 "Hank Schra… <chr [1]>             62 Dean No… 415192 dean-norris  nm06… 14329
#> 5 "Marie Schr… <chr [1]>             62 Betsy B… 436547 betsy-brandt nm13… 1217…
#> 6 "Walter Whi… <chr [1]>             62 RJ Mitte 459150 rj-mitte-06… nm26… 2096…
#> 7 "Saul Goodm… <chr [1]>             42 Bob Ode… 294857 bob-odenkirk nm06… 59410
#> 8 "Mike Ehrma… <chr [1]>             42 Jonatha… 411251 jonathan-ba… nm00… 783  
#> 9 "Gustavo \"… <chr [1]>             26 Giancar… 412546 giancarlo-e… nm00… 4808 
#> 
#> $crew
#> # A tibble: 61 x 9
#>    job      jobs   episode_count name      trakt  slug     imdb  tmdb  crew_type
#>    <chr>    <list>         <int> <chr>     <chr>  <chr>    <chr> <chr> <chr>    
#>  1 Executi… <chr …            62 Michelle… 12427  michell… nm05… 29779 producti…
#>  2 Producer <chr …            62 Peter Go… 20811  peter-g… nm03… 24951 producti…
#>  3 Producer <chr …            62 Thomas S… 67528  thomas-… nm10… 1030… producti…
#>  4 Producer <chr …            62 Bryan Cr… 297737 bryan-c… nm01… 17419 producti…
#>  5 Producer <chr …            62 Stewart … 325805 stewart… nm05… 1223… producti…
#>  6 Producer <chr …            62 Moira Wa… 325808 moira-w… nm09… 1223… producti…
#>  7 Producer <chr …            62 Karen Mo… 365840 karen-m… nm10… 1223… producti…
#>  8 Producer <chr …            62 Patty Lin 398571 patty-l… nm05… 1218… producti…
#>  9 Producer <chr …            62 Melissa … 400288 melissa… nm22… 1223… producti…
#> 10 Producer <chr …            62 George M… 400289 george-… nm22… 1223… producti…
#> # … with 51 more rows
#> 

seasons_people("breaking-bad", season = 1)

#> $cast
#> # A tibble: 6 x 8
#>   character   characters episode_count name    trakt  slug           imdb  tmdb 
#>   <chr>       <list>             <int> <chr>   <chr>  <chr>          <chr> <chr>
#> 1 Walter Whi… <chr [1]>              7 Bryan … 297737 bryan-cranston nm01… 17419
#> 2 Jesse Pink… <chr [1]>              7 Aaron … 415249 aaron-paul     nm06… 84497
#> 3 Skyler Whi… <chr [1]>              7 Anna G… 420986 anna-gunn      nm03… 1345…
#> 4 Hank Schra… <chr [1]>              7 Dean N… 415192 dean-norris    nm06… 14329
#> 5 Marie Schr… <chr [1]>              7 Betsy … 436547 betsy-brandt   nm13… 1217…
#> 6 Walter Whi… <chr [1]>              7 RJ Mit… 459150 rj-mitte-06c3… nm26… 2096…
#> 
#> $crew
#> # A tibble: 30 x 9
#>    job      jobs   episode_count name      trakt  slug     imdb  tmdb  crew_type
#>    <chr>    <list>         <int> <chr>     <chr>  <chr>    <chr> <chr> <chr>    
#>  1 Executi… <chr …             7 Michelle… 12427  michell… nm05… 29779 producti…
#>  2 Producer <chr …             7 Peter Go… 20811  peter-g… nm03… 24951 producti…
#>  3 Producer <chr …             7 Thomas S… 67528  thomas-… nm10… 1030… producti…
#>  4 Producer <chr …             7 Bryan Cr… 297737 bryan-c… nm01… 17419 producti…
#>  5 Producer <chr …             7 Stewart … 325805 stewart… nm05… 1223… producti…
#>  6 Producer <chr …             7 Moira Wa… 325808 moira-w… nm09… 1223… producti…
#>  7 Producer <chr …             7 Karen Mo… 365840 karen-m… nm10… 1223… producti…
#>  8 Producer <chr …             7 Patty Lin 398571 patty-l… nm05… 1218… producti…
#>  9 Producer <chr …             7 Melissa … 400288 melissa… nm22… 1223… producti…
#> 10 Producer <chr …             7 George M… 400289 george-… nm22… 1223… producti…
#> # … with 20 more rows
#> 

episodes_people("breaking-bad", season = 1, episode = 1)

#> $cast
#> # A tibble: 6 x 7
#>   character     characters name      trakt  slug                    imdb   tmdb 
#>   <chr>         <list>     <chr>     <chr>  <chr>                   <chr>  <chr>
#> 1 Walter White  <chr [1]>  Bryan Cr… 297737 bryan-cranston          nm018… 17419
#> 2 Jesse Pinkman <chr [1]>  Aaron Pa… 415249 aaron-paul              nm066… 84497
#> 3 Skyler White  <chr [1]>  Anna Gunn 420986 anna-gunn               nm034… 1345…
#> 4 Hank Schrader <chr [1]>  Dean Nor… 415192 dean-norris             nm060… 14329
#> 5 Marie Schrad… <chr [1]>  Betsy Br… 436547 betsy-brandt            nm133… 1217…
#> 6 Walter White… <chr [1]>  RJ Mitte  459150 rj-mitte-06c377f6-88c1… nm266… 2096…
#> 
#> $crew
#> # A tibble: 3 x 8
#>   job              jobs     name        trakt slug       imdb    tmdb  crew_type
#>   <chr>            <list>   <chr>       <chr> <chr>      <chr>   <chr> <chr>    
#> 1 Director         <chr [1… Vince Gill… 4230… vince-gil… nm0319… 66633 directing
#> 2 Writer           <chr [1… Vince Gill… 4230… vince-gil… nm0319… 66633 writing  
#> 3 Director of Pho… <chr [1… John Toll   2172  john-toll  nm0001… 2483  camera   
#> 

# }