Get the cast and crew of a show or movie

Returns all cast and crew for a show/movie, depending on how much data is available.

movies_people(id, extended = c("min", "full"))

shows_people(id, guest_stars = FALSE, extended = c("min", "full"))

seasons_people(
  id,
  season = 1L,
  guest_stars = FALSE,
  extended = c("min", "full")
)

episodes_people(
  id,
  season = 1L,
  episode = 1L,
  guest_stars = FALSE,
  extended = c("min", "full")
)

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things") for more details.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things") for more details.

guest_stars
logical(1) ["FALSE"]: Also include guest stars. This returns a lot of data, so use with care.

season
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

episode
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Source

movies_people() wraps endpoint movies/:id/people.

shows_people() wraps endpoint shows/:id/people.

seasons_people() wraps endpoint shows/:id/seasons/:season/people.

episodes_people() wraps endpoint shows/:id/seasons/:season/episodes/:episode/people.

Value

A list of one or more tibbles for cast and/or crew. The latter tibble objects are as flat as possible.

Note

As of 2019-09-30, there are two representations of character[s] and job[s]: One is a regular character variable, and the other is a list-column. The former is deprecated and only included for compatibility reasons.

See also

Examples

if (FALSE) {
movies_people("deadpool-2016")
shows_people("breaking-bad")
seasons_people("breaking-bad", season = 1)
episodes_people("breaking-bad", season = 1, episode = 1)
}