Get the cast and crew of a show or movie

Returns all cast and crew for a show/movie, depending on how much data is available.

trakt.shows.people(target, extended = c("min", "full"))

trakt.movies.people(target, extended = c("min", "full"))

Arguments

target
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all target values separately and the result is combined.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

Value

A list of one or more tibbles for cast and crew. The latter tibble objects are as flat as possible.

Details

The API methods for these functions are:

/shows/:id/people

/movies/:id/people

Note that as of 2019-09-30, there are two representations of character[s] and job[s]: One is a regular character variable, and the other is a list-column. The former is deprecated and only included for compatibility reasons.

See also

Examples

# \dontrun{
trakt.shows.people("breaking-bad")

#> $cast
#> # A tibble: 9 x 8
#>   character    characters episode_count name     trakt  slug         imdb  tmdb 
#>   <chr>        <list>             <int> <chr>    <chr>  <chr>        <chr> <chr>
#> 1 Walter White <chr [1]>             62 Bryan C… 297737 bryan-crans… nm01… 17419
#> 2 Jesse Pinkm… <chr [1]>             62 Aaron P… 415249 aaron-paul   nm06… 84497
#> 3 Skyler White <chr [1]>             62 Anna Gu… 420986 anna-gunn    nm03… 1345…
#> 4 Hank Schrad… <chr [1]>             62 Dean No… 415192 dean-norris  nm06… 14329
#> 5 Marie Schra… <chr [1]>             62 Betsy B… 436547 betsy-brandt nm13… 1217…
#> 6 Walter Whit… <chr [1]>             62 RJ Mitte 459150 rj-mitte-06… nm26… 2096…
#> 7 Saul Goodman <chr [1]>             42 Bob Ode… 294857 bob-odenkirk nm06… 59410
#> 8 Mike Ehrman… <chr [1]>             42 Jonatha… 411251 jonathan-ba… nm00… 783  
#> 9 "Gustavo \"… <chr [1]>             26 Giancar… 412546 giancarlo-e… nm00… 4808 
#> 
#> $crew
#> # A tibble: 61 x 9
#>    job      jobs   episode_count name      trakt  slug     imdb  tmdb  crew_type
#>    <chr>    <list>         <int> <chr>     <chr>  <chr>    <chr> <chr> <chr>    
#>  1 Executi… <chr …            62 Michelle… 12427  michell… nm05… 29779 producti…
#>  2 Producer <chr …            62 Peter Go… 20811  peter-g… nm03… 24951 producti…
#>  3 Producer <chr …            62 Thomas S… 67528  thomas-… nm10… 1030… producti…
#>  4 Producer <chr …            62 Bryan Cr… 297737 bryan-c… nm01… 17419 producti…
#>  5 Producer <chr …            62 Stewart … 325805 stewart… nm05… 1223… producti…
#>  6 Producer <chr …            62 Moira Wa… 325808 moira-w… nm09… 1223… producti…
#>  7 Producer <chr …            62 Karen Mo… 365840 karen-m… nm10… 1223… producti…
#>  8 Producer <chr …            62 Patty Lin 398571 patty-l… nm05… 1218… producti…
#>  9 Producer <chr …            62 Melissa … 400288 melissa… nm22… 1223… producti…
#> 10 Producer <chr …            62 George M… 400289 george-… nm22… 1223… producti…
#> # … with 51 more rows
#> 

# }