Media user ratings

Returns a movie's or show's (or season's, or episode's) rating and ratings distribution. If you do not want the full ratings distribution, it is highly advised to just use *_summary functions or seasons_season for episode ratings.

shows_ratings(id)

movies_ratings(id)

seasons_ratings(id, season = 1L)

episodes_ratings(id, season = 1L, episode = 1L)

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things", package = "tRakt") for more details.

season
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

episode
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Source

shows_ratings() wraps endpoint shows/:ids/ratings.

movies_ratings() wraps endpoint movies/:id/ratings.

seasons_ratings() wraps endpoint shows/:id/seasons/:season/ratings.

episodes_ratings() wraps endpoint shows/:id/seasons/:season/episodes/:episode/ratings.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

Note

Since this function is able to work on multi-length inputs for id, season and episode, it is possible to get a lot of data, but at the cost of one API call per element in each argument. Please be kind to the API.

See also

Examples

# A movie's ratings
movies_ratings("tron-legacy-2010")

#> # A tibble: 1 x 5
#>   rating votes distribution      id               type  
#>    <dbl> <int> <list>            <chr>            <chr> 
#> 1   7.27 13193 <tibble [10 × 2]> tron-legacy-2010 movies

# A show's ratings
shows_ratings("game-of-thrones")

#> # A tibble: 1 x 5
#>   rating  votes distribution      id              type 
#>    <dbl>  <int> <list>            <chr>           <chr>
#> 1   9.19 105723 <tibble [10 × 2]> game-of-thrones shows

# \dontrun{
# Ratings for seasons 1 through 5
seasons_ratings("futurama", season = 1:5)

#> # A tibble: 5 x 5
#>   rating votes distribution      id       season
#>    <dbl> <int> <list>            <chr>     <int>
#> 1   8.23   252 <tibble [10 × 2]> futurama      1
#> 2   8.25   188 <tibble [10 × 2]> futurama      2
#> 3   8.44   162 <tibble [10 × 2]> futurama      3
#> 4   8.26   154 <tibble [10 × 2]> futurama      4
#> 5   8.23   129 <tibble [10 × 2]> futurama      5

# Ratings for episodes 1 through 7 of season 1
episodes_ratings("futurama", season = 1, episode = 1:7)

#> # A tibble: 7 x 6
#>   rating votes distribution      id       season episode
#>    <dbl> <int> <list>            <chr>     <int>   <int>
#> 1   8.11  2097 <tibble [10 × 2]> futurama      1       1
#> 2   7.87  1635 <tibble [10 × 2]> futurama      1       2
#> 3   7.88  1496 <tibble [10 × 2]> futurama      1       3
#> 4   7.79  1341 <tibble [10 × 2]> futurama      1       4
#> 5   7.77  1292 <tibble [10 × 2]> futurama      1       5
#> 6   7.83  1263 <tibble [10 × 2]> futurama      1       6
#> 7   7.75  1176 <tibble [10 × 2]> futurama      1       7

# }