Get a show or movie's (or season's or episode's) stats

The data contains watchers, playes, collectors, comments, lists, and votes.

shows_stats(id)

movies_stats(id)

seasons_stats(id, season = 1L)

episodes_stats(id, season = 1L, episode = 1L)

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things", package = "tRakt") for more details.

season
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

episode
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Source

shows_stats() wraps endpoint shows/:id/stats.

movies_stats() wraps endpoint movies/:id/stats.

seasons_stats() wraps endpoint shows/:id/seasons/:season/stats.

episodes_stats() wraps endpoint shows/:id/seasons/:season/episodes/:episode/stats.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

See also

Examples

# Stats for a movie
movies_stats("inception-2010")

#> # A tibble: 1 x 8
#>   watchers  plays collectors comments lists votes type   id            
#>      <int>  <int>      <int>    <int> <int> <int> <chr>  <chr>         
#> 1   323955 431940     229207      120 46800 55762 movies inception-2010

if (FALSE) {
# Stats for multiple shows at once
shows_stats(c("breaking-bad", "game-of-thrones"))

# Stats for multiple episodes
episodes_stats("futurama", season = 1, episode = 1:7)
}