Get a show or movie's stats

The data contains watchers, playes, collectors, comments, lists, and votes.

trakt.shows.stats(target)

trakt.movies.stats(target)

trakt.seasons.stats(target, season = 1L)

trakt.episodes.stats(target, season = 1L, episode = 1L)

Arguments

target
character(1): The ID of the item requested. Preferably the Trakt ID (e.g. 1390), slug (e.g. "game-of-thrones"), or IMDb ID (e.g. "tt0944947"). Can also be of length greater than 1, in which case the function is called on all target values separately and the result is combined.

season
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

episode
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Value

A tibble().

Examples

# Stats for a movie
trakt.movies.stats("inception-2010")

#> # A tibble: 1 x 8
#>   watchers  plays collectors comments lists votes type   id            
#>      <int>  <int>      <int>    <int> <int> <int> <chr>  <chr>         
#> 1   254749 339666     156115      106 38024 47577 movies inception-2010

# Stats for multiple shows at once
trakt.shows.stats(c("breaking-bad", "game-of-thrones"))

#> # A tibble: 2 x 9
#>   watchers   plays collectors collected_episo… comments  lists votes type  id   
#>      <int>   <int>      <int>            <int>    <int>  <int> <int> <chr> <chr>
#> 1   418244  2.55e7     345554         17327367      403 119302 64040 shows brea…
#> 2   773373  4.57e7    1104336         58519066      331 263881 88298 shows game…

# Stats for multiple episodes
trakt.episodes.stats("futurama", season = 1, episode = 1:7)

#> # A tibble: 7 x 9
#>   watchers  plays collectors comments lists votes season episode id      
#>      <int>  <int>      <int>    <int> <int> <int>  <dbl>   <int> <chr>   
#> 1    87280 116374     115886        1   171  1914      1       1 futurama
#> 2    83152 110870     115422        1   116  1498      1       2 futurama
#> 3    82050 109400     115410        0   103  1363      1       3 futurama
#> 4    80516 107355     114825        0   106  1215      1       4 futurama
#> 5    79902 106536     115053        1   104  1169      1       5 futurama
#> 6    79798 106398     114725        0   103  1153      1       6 futurama
#> 7    79019 105359     114645        0   102  1070      1       7 futurama