Get a show or movie's stats

The data contains watchers, playes, collectors, comments, lists, and votes.

trakt.shows.stats(target)

trakt.movies.stats(target)

trakt.seasons.stats(target, season = 1L)

trakt.episodes.stats(target, season = 1L, episode = 1L)

Arguments

target
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all target values separately and the result is combined.

season
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

episode
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Value

A tibble().

Details

The API methods for these functions are:

/movies/:id/stats

/shows/:id/stats

/shows/:id/seasons/:season/stats

/shows/:id/seasons/:season/episodes/:episode/stats

Examples

# Stats for a movie
trakt.movies.stats("inception-2010")

#> # A tibble: 1 x 8
#>   watchers  plays collectors comments lists votes type   id            
#>      <int>  <int>      <int>    <int> <int> <int> <chr>  <chr>         
#> 1   286670 382227     188891      114 42363 51474 movies inception-2010

# \dontrun{
# Stats for multiple shows at once
trakt.shows.stats(c("breaking-bad", "game-of-thrones"))

#> # A tibble: 2 x 9
#>   watchers   plays collectors collected_episo… comments  lists votes type  id   
#>      <int>   <int>      <int>            <int>    <int>  <int> <int> <chr> <chr>
#> 1   468228  2.85e7     405265         20149833      425 130097 69153 shows brea…
#> 2   916399  5.57e7    1291030         69853992      342 271078 99025 shows game…

# Stats for multiple episodes
trakt.episodes.stats("futurama", season = 1, episode = 1:7)

#> # A tibble: 7 x 9
#>   watchers  plays collectors comments lists votes season episode id      
#>      <int>  <int>      <int>    <int> <int> <int>  <dbl>   <int> <chr>   
#> 1    97624 130166     142013        2   177  2032      1       1 futurama
#> 2    92538 123384     141536        1   116  1585      1       2 futurama
#> 3    91313 121751     141566        0   107  1453      1       3 futurama
#> 4    90007 120010     140958        0   108  1294      1       4 futurama
#> 5    88921 118562     141185        1   108  1251      1       5 futurama
#> 6    88656 118209     140842        0   105  1228      1       6 futurama
#> 7    87700 116934     140698        0   104  1146      1       7 futurama

# }