Get who's watching a thing right now

Get who's watching a thing right now

movies_watching(id, extended = c("min", "full"))

shows_watching(id, extended = c("min", "full"))

seasons_watching(id, season = 1L, extended = c("min", "full"))

episodes_watching(id, season = 1L, episode = 1L, extended = c("min", "full"))

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things", package = "tRakt") for more details.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things", package = "tRakt") for more details.

season
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

episode
integer(1) [1L]: The season and eisode number. If longer, e.g. 1:5, the function is vectorized and the output will be combined. This may result in a lot of API calls. Use wisely.

Source

movies_watching() wraps endpoint movies/:id/watching.

shows_watching() wraps endpoint shows/:id/watching.

episodes_watching() wraps endpoint shows/:id/seasons/:season/episodes/:episode/watching.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

Functions

movies_watching: Who's watching a movie.

shows_watching: Who's watching a show.

seasons_watching: Who's watching a season.

episodes_watching: Who's watching an episode.

See also

Examples

# \dontrun{
movies_watching("deadpool-2016")

#> # A tibble: 1 x 6
#>   username private user_name vip   vip_ep user_slug
#>   <chr>    <lgl>   <lgl>     <lgl> <lgl>  <chr>    
#> 1 flagg    FALSE   NA        FALSE FALSE  flagg    

shows_watching("the-simpsons")

#> # A tibble: 23 x 6
#>    username            private user_name        vip   vip_ep user_slug          
#>    <chr>               <lgl>   <chr>            <lgl> <lgl>  <chr>              
#>  1 orthodoxmonster     FALSE   <NA>             FALSE FALSE  orthodoxmonster    
#>  2 filipve             FALSE   Filip Vanden Ey… FALSE FALSE  filipve            
#>  3 bublinak            TRUE    <NA>             NA    NA     bublinak           
#>  4 DashRendar          FALSE   Dashiel MacTavi… FALSE FALSE  dashrendar         
#>  5 Guyde               TRUE    <NA>             NA    NA     guyde              
#>  6 tomislav_sb         FALSE   <NA>             FALSE FALSE  tomislav_sb        
#>  7 deusexmachina354    FALSE   Tyler            FALSE FALSE  deusexmachina354   
#>  8 papper              FALSE   Papper           FALSE FALSE  papper             
#>  9 moreton.andy@gmail… FALSE   Andy             FALSE FALSE  moreton-andy-gmail…
#> 10 jackyaz             TRUE    <NA>             NA    NA     jackyaz            
#> # … with 13 more rows

seasons_watching("the-simpsons", season = 6)

#> # A tibble: 2 x 6
#>   username    private user_name vip   vip_ep user_slug  
#>   <chr>       <lgl>   <lgl>     <lgl> <lgl>  <chr>      
#> 1 Guyde       TRUE    NA        NA    NA     guyde      
#> 2 tomislav_sb FALSE   NA        FALSE FALSE  tomislav_sb

episodes_watching("the-simpsons", season = 6, episode = 12)

#> # A tibble: 0 x 0

# }