Get a single person's movie or show credits

Returns all movies or shows where this person is in the cast or crew.

people_movies(id, extended = c("min", "full"))

people_shows(id, extended = c("min", "full"))

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things") for more details.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things") for more details.

Source

people_movies() wraps endpoint people/:id/movies.

people_shows() wraps endpoint people/:id/shows.

Value

A list of one or more tibbles for cast and crew. The latter tibble objects are as flat as possible.

Details

Note that as of 2019-09-30, there are two representations of character[s] and job[s]: One is a regular character variable, and the other is a list-column. The singular is deprecated and only included for compatibility reasons.

See also

Examples

if (FALSE) {
people_movies("christopher-nolan")

people_shows("kit-harington")
}