Get a single person's movie or show credits

Returns all movies or shows where this person is in the cast or crew.

people_movies(id, extended = c("min", "full"))

people_shows(id, extended = c("min", "full"))

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things", package = "tRakt") for more details.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things", package = "tRakt") for more details.

Source

people/:id/movies

people/:id/shows

Value

A list of one or more tibbles for cast and crew. The latter tibble objects are as flat as possible.

Details

Note that as of 2019-09-30, there are two representations of character[s] and job[s]: One is a regular character variable, and the other is a list-column. The singular is deprecated and only included for compatibility reasons.

See also

Examples

# \dontrun{
people_movies("christopher-nolan")

#> $cast
#> # A tibble: 17 x 8
#>    character characters title            year trakt  slug            imdb  tmdb 
#>    <chr>     <list>     <chr>           <int> <chr>  <chr>           <chr> <chr>
#>  1 Himself   <chr [1]>  Making Waves: …  2019 441872 making-waves-t… tt38… 5958…
#>  2 himself   <chr [1]>  The Making of …  2017 399062 the-making-of-… tt77… 5491…
#>  3 ""        <chr [0]>  Cinema Futures   2016 285889 cinema-futures… tt60… 4168…
#>  4 ""        <chr [0]>  Inside Interst…  2015 378906 inside-interst… <NA>  5291…
#>  5 Himself   <chr [1]>  Interstellar: …  2014 193834 interstellar-n… tt41… 3019…
#>  6 Himself   <chr [1]>  Christopher No…  2013 147862 christopher-no… tt31… 2428…
#>  7 Himself   <chr [1]>  The Fire Rises…  2013 148113 the-fire-rises… tt31… 2432…
#>  8 Himself   <chr [1]>  Tales from the…  2013 139154 tales-from-the… tt25… 2275…
#>  9 Himself   <chr [1]>  Side by Side     2012 77846  side-by-side-2… tt20… 1103…
#> 10 Himself   <chr [1]>  The Batmobile    2012 86120  the-batmobile-… tt22… 1252…
#> 11 Himself   <chr [1]>  Exploring The …  2011 189583 exploring-the-… tt21… 2927…
#> 12 himself   <chr [1]>  These Amazing …  2011 59813  these-amazing-… tt12… 80337
#> 13 self      <chr [1]>  Inception: Jum…  2010 459731 inception-jump… tt52… 6164…
#> 14 Himself   <chr [1]>  Batman Tech      2008 116883 batman-tech-20… tt12… 1865…
#> 15 himself   <chr [1]>  Batman Unmaske…  2008 18787  batman-unmaske… tt12… 29751
#> 16 Himself   <chr [1]>  Gotham Uncover…  2008 116876 gotham-uncover… <NA>  1865…
#> 17 Himself   <chr [1]>  Film Noir: Bri…  2006 216490 film-noir-brin… tt08… 3119…
#> 
#> $crew
#> # A tibble: 57 x 9
#>    job      jobs   title          year trakt slug          imdb  tmdb  crew_type
#>    <chr>    <list> <chr>         <int> <chr> <chr>         <chr> <chr> <chr>    
#>  1 Producer <chr … Tenet          2020 4279… tenet-2020    tt67… 5779… producti…
#>  2 Executi… <chr … Justice Leag…  2017 94232 justice-leag… tt09… 1410… producti…
#>  3 Producer <chr … Dunkirk        2017 2246… dunkirk-2017  tt50… 3747… producti…
#>  4 Executi… <chr … Batman v Sup…  2016 2489… batman-v-sup… <NA>  4042… producti…
#>  5 Executi… <chr … Batman v Sup…  2016 1295… batman-v-sup… tt29… 2091… producti…
#>  6 Producer <chr … Quay           2015 2470… quay-2015     tt49… 3521… producti…
#>  7 Producer <chr … Interstellar   2014 1021… interstellar… tt08… 1573… producti…
#>  8 Executi… <chr … Transcendence  2014 1021… transcendenc… tt22… 1573… producti…
#>  9 Producer <chr … Man of Steel   2013 34507 man-of-steel… tt07… 49521 producti…
#> 10 Producer <chr … The Dark Kni…  2012 34073 the-dark-kni… tt13… 49026 producti…
#> # … with 47 more rows
#> 

people_shows("kit-harington")

#> $cast
#> # A tibble: 3 x 11
#>   title  year trakt slug  tvdb  imdb  tmdb  character characters episode_count
#>   <chr> <int> <chr> <chr> <chr> <chr> <chr> <chr>     <list>             <int>
#> 1 Gunp…  2017 1232… gunp… 3340… tt61… 73550 Robert C… <chr [1]>              3
#> 2 Game…  2011 1390  game… 1213… tt09… 1399  Jon Snow  <chr [1]>             73
#> 3 Satu…  1975 1656  satu… 76177 tt00… 1667  Himself … <chr [1]>              2
#> # … with 1 more variable: series_regular <lgl>
#> 

# }