Get a single person's movie or show credits

Returns all movies or shows where this person is in the cast or crew.

trakt.people.movies(target, extended = c("min", "full"))

trakt.people.shows(target, extended = c("min", "full"))

Arguments

target
character(1): The ID of the item requested. Preferably the Trakt ID (e.g. 1390), slug (e.g. "game-of-thrones"), or IMDb ID (e.g. "tt0944947"). Can also be of length greater than 1, in which case the function is called on all target values separately and the result is combined.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

Value

A list of one or more tibbles for cast and crew. The latter tibble objects are as flat as possible.

See also

Examples

trakt.people.movies("christopher-nolan")

#> $cast
#> # A tibble: 15 x 7
#>    character title                 year trakt  slug                  imdb  tmdb 
#>    <chr>     <chr>                <int> <chr>  <chr>                 <chr> <chr>
#>  1 himself   The Making of Dunki…  2017 399062 the-making-of-dunkir… tt77… 5491…
#>  2 ""        Cinema Futures        2016 285889 cinema-futures-2016   tt60… 4168…
#>  3 ""        Inside Interstellar   2015 378906 inside-interstellar-… <NA>  5291…
#>  4 Himself   Interstellar: Nolan…  2014 193834 interstellar-nolan-s… tt41… 3019…
#>  5 Himself   Tales from the Warn…  2013 139154 tales-from-the-warne… tt25… 2275…
#>  6 Himself   Christopher Nolan &…  2013 147862 christopher-nolan-ri… tt31… 2428…
#>  7 Himself   The Fire Rises: The…  2013 148113 the-fire-rises-the-c… tt31… 2432…
#>  8 Himself   Side by Side          2012 77846  side-by-side-2012     tt20… 1103…
#>  9 Himself   The Batmobile         2012 86120  the-batmobile-2012    tt22… 1252…
#> 10 himself   These Amazing Shado…  2011 59813  these-amazing-shadow… tt12… 80337
#> 11 Himself   Exploring The Tree …  2011 189583 exploring-the-tree-o… tt21… 2927…
#> 12 himself   Batman Unmasked: Th…  2008 18787  batman-unmasked-the-… tt12… 29751
#> 13 Himself   Gotham Uncovered: C…  2008 116876 gotham-uncovered-cre… <NA>  1865…
#> 14 Himself   Batman Tech           2008 116883 batman-tech-2008      tt12… 1865…
#> 15 Himself   Film Noir: Bringing…  2006 216490 film-noir-bringing-d… tt08… 3119…
#> 
#> $crew
#> # A tibble: 55 x 8
#>    job      title             year trakt  slug             imdb  tmdb  crew_type
#>    <chr>    <chr>            <int> <chr>  <chr>            <chr> <chr> <chr>    
#>  1 Executi… Justice League    2017 94232  justice-league-… tt09… 1410… producti…
#>  2 Producer Dunkirk           2017 224641 dunkirk-2017     tt50… 3747… producti…
#>  3 Executi… Batman v Superm…  2016 129583 batman-v-superm… tt29… 2091… producti…
#>  4 Executi… Batman v Superm…  2016 248921 batman-v-superm… <NA>  4042… producti…
#>  5 Producer Quay              2015 247092 quay-2015        tt49… 3521… producti…
#>  6 Producer Interstellar      2014 102156 interstellar-20… tt08… 1573… producti…
#>  7 Executi… Transcendence     2014 102164 transcendence-2… tt22… 1573… producti…
#>  8 Producer Man of Steel      2013 34507  man-of-steel-20… tt07… 49521 producti…
#>  9 Producer The Dark Knight…  2012 34073  the-dark-knight… tt13… 49026 producti…
#> 10 Producer Inception         2010 16662  inception-2010   tt13… 27205 producti…
#> # … with 45 more rows
#> 

trakt.people.shows("kit-harington")

#> $cast
#> # A tibble: 2 x 8
#>   title           year trakt  slug           tvdb   imdb     tmdb  character    
#>   <chr>          <int> <chr>  <chr>          <chr>  <chr>    <chr> <chr>        
#> 1 Gunpowder       2017 123241 gunpowder      334069 tt61282… 73550 Robert Cates…
#> 2 Game of Thron…  2011 1390   game-of-thron… 121361 tt09449… 1399  Jon Snow     
#>