Popular media

These functions return the popular movies/shows on trakt.tv.

movies_popular(limit = 10, extended = c("min", "full"), query = NULL,
  years = NULL, genres = NULL, languages = NULL, countries = NULL,
  runtimes = NULL, ratings = NULL, certifications = NULL)

shows_popular(limit = 10, extended = c("min", "full"), query = NULL,
  years = NULL, genres = NULL, languages = NULL, countries = NULL,
  runtimes = NULL, ratings = NULL, certifications = NULL,
  networks = NULL, status = NULL)

Arguments

limit
integer(1) [10L]: Number of items to return. Must be greater than 0 and will be coerced via as.integer().

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things", package = "tRakt") for more details.

query
character(1): Search string for titles and descriptions. For search_query() other fields are searched depending on the type of media. See the API docs for a full reference.

years
character | integer: 4-digit year (2010) or range, e.g. "2010-2020". Can also be an integer vector of length two which will be coerced appropriately, e.g. c(2010, 2020).

genres
character(n): Genre slug(s). See trakt_genres for a table of genres. Multiple values are allowed and will be concatenated.

languages
character(n): Two-letter language code(s). Also see trakt_languages for available languages (code and name).

countries
character(n): Two-letter country code(s). See trakt_countries.

runtimes
character | integer: Integer range in minutes, e.g. 30-90. Can also be an integer vector of length two which will be coerced appropriately.

ratings
character | integer: Integer range between 0 and 100. Can also be an integer vector of length two which will be coerced appropriately. Note that user-supplied ratings are in the range of 1 to 10, yet the ratings on the site itself are scaled to the range of 1 to 100.

certifications
character(n): Certification(s) like pg-13. Multiple values are allowed. Use trakt_certifications for reference. Note that there are different certifications for shows and movies.

networks
character(n): (Shows only) Network name like HBO. See trakt_networks for a list of known networks.

status
character(n): (Shows only) The status of the shows. One of "returning series", "in production", "planned", "canceled", or "ended".

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

The Dynamic Lists on trakt.tv

These functions access the automatically updated lists provided by trakt.tv. Each function comes in two flavors: Shows or movies. The following descriptions are adapted directly from the API reference.

Popular: Popularity is calculated using the rating percentage and the number of ratings.

Trending: Returns all movies/shows being watched right now. Movies/shows with the most users are returned first.

Played: Returns the most played (a single user can watch multiple times) movies/shows in the specified time period.

Watched: Returns the most watched (unique users) movies/shows in the specified time period.

Collected: Returns the most collected (unique users) movies/shows in the specified time period.

Anticipated: Returns the most anticipated movies/shows based on the number of lists a movie/show appears on.

Updates: Returns all movies/shows updated since the specified UTC start_date. In this case, the upper bound for limit is 100.

The functions for Played, Watched, Collected and Played each return the same additional variables besides the media information: watcher_count, play_count, collected_count, collector_count.

See also

Examples

# \dontrun{
# Get the most popular German-language movies between 2000 and 2010
movies_popular(languages = "de", years = c(2000, 2010))

#> # A tibble: 10 x 6
#>    title                    year trakt slug                       imdb      tmdb
#>    <chr>                   <int> <int> <chr>                      <chr>    <int>
#>  1 Downfall                 2004   489 downfall-2004              tt03631…   613
#>  2 The Lives of Others      2006   461 the-lives-of-others-2006   tt04050…   582
#>  3 The Wave                 2008  3954 the-wave-2008              tt10636…  7735
#>  4 Good bye, Lenin!         2003   285 good-bye-lenin-2003        tt03013…   338
#>  5 The Experiment           2001   454 the-experiment-2001        tt02502…   575
#>  6 The White Ribbon         2009 24371 the-white-ribbon-2009      tt11493… 37903
#>  7 The Baader Meinhof Com…  2008  3789 the-baader-meinhof-comple… tt07654…  6968
#>  8 Cargo                    2009 21807 cargo-2009                 tt03819… 34069
#>  9 The Counterfeiters       2007  3967 the-counterfeiters-2007    tt08135…  7862
#> 10 Manitou's Shoe           2001  4152 manitou-s-shoe-2001        tt02484…  8366

# }