Get a season of a show

Similar to seasons_summary, but this function returns full data for a single season, i.e. all the episodes of the season

Usage

seasons_season(id, seasons = 1L, extended = c("min", "full"))

Source

seasons_season() wraps endpoint shows/id/seasons/season.

Arguments

id

character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things") for more details.

seasons

integer(1) [1L]: The season(s) to get. Use 0 for specials.

extended

character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things") for more details.

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

Note

If you want to quickly gather episode data of all available seasons, see seasons_summary and use the episodes = TRUE parameter.

See also

Examples

if (FALSE) {
seasons_season("breaking-bad", 1)

# Including all episode data:
seasons_season("breaking-bad", 1, extended = "full")
}