Get a season of a show

Similar to seasons_summary, but this function returns full data for a single season, i.e. all the episodes of the season

seasons_season(id, seasons = 1L, extended = c("min", "full"))

Arguments

id
character(1): The ID of the item requested. Preferably the trakt ID (e.g. 1429). Other options are the trakt.tv slug (e.g. "the-wire") or imdb ID (e.g. "tt0306414"). Can also be of length greater than 1, in which case the function is called on all id values separately and the result is combined. See vignette("finding-things", package = "tRakt") for more details.

seasons
integer(1) [1L]: The season(s) to get. Use 0 for specials.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required. See vignette("finding-things", package = "tRakt") for more details.

Source

shows/id/seasons/season

Value

A tibble(). If the function has a limit parameter (defaulting to 10), this will be the (maximum) number of rows of the tibble. If there are no results (or the API is unreachable), an empty tibble() is returned.

Note

If you want to quickly gather episode data of all available seasons, see seasons_summary and use the episodes = TRUE parameter.

See also

Examples

# \dontrun{
seasons_season("breaking-bad", 1)

#> # A tibble: 7 x 7
#>   season episode title                         trakt tvdb   imdb      tmdb 
#>    <int>   <int> <chr>                         <chr> <chr>  <chr>     <chr>
#> 1      1       1 Pilot                         73482 349232 tt0959621 62085
#> 2      1       2 Cat's in the Bag...           73483 349233 tt1054724 62086
#> 3      1       3 ...And the Bag's in the River 73484 349235 tt1054725 62087
#> 4      1       4 Cancer Man                    73485 349236 tt1054726 62088
#> 5      1       5 Gray Matter                   73486 349238 tt1054727 62089
#> 6      1       6 Crazy Handful of Nothin'      73487 355100 tt1054728 62090
#> 7      1       7 A No-Rough-Stuff-Type Deal    73488 352534 tt1054729 62091

# Including all episode data:
seasons_season("breaking-bad", 1, extended = "full")

#> # A tibble: 7 x 16
#>   season episode title episode_abs overview rating votes comment_count
#>    <int>   <int> <chr>       <int> <chr>     <dbl> <int>         <int>
#> 1      1       1 Pilot           1 When an…   8.35  7103            26
#> 2      1       2 Cat'…           2 Walt an…   8.14  5600             3
#> 3      1       3 ...A…           3 Walter …   8.06  5027             2
#> 4      1       4 Canc…           4 Walter …   8.00  4868             3
#> 5      1       5 Gray…           5 Walter …   7.99  4752             3
#> 6      1       6 Craz…           6 The sid…   8.52  4842             3
#> 7      1       7 A No…           7 Walter …   8.40  4699             2
#> # … with 8 more variables: first_aired <dttm>, updated_at <dttm>,
#> #   available_translations <list>, runtime <int>, trakt <chr>, tvdb <chr>,
#> #   imdb <chr>, tmdb <chr>

# }