tRakt: Get Data from 'trakt.tv'

A wrapper for the https://trakt.tv API to retrieve data about shows and movies, including user ratings, credits and related metadata. Additional functions retrieve user-specific information including collections and history of watched items. A full API reference is available at https://trakt.docs.apiary.io.

See also

Author

Maintainer: Lukas Burk lukas@quantenbrot.de (ORCID)