Get a show's episodes. All of them.

trakt.get_all_episodes pulls detailed episode data. Get details for a show's episodes, e.g. ratings, number of votes, airdates, images, plot overviews…

trakt.get_all_episodes(target, season_nums = NULL, extended = "full",
  dropunaired = TRUE)

Arguments

target
The id of the show requested. Either the slug (e.g. "game-of-thrones"), trakt id or IMDb id.

season_nums
Vector of season numbers, e.g. c(1, 2). If NULL, all the seasons are pulled by calling trakt.seasons.summary to determine the number of seasons. If a vector of length 1 is supplied, it is extended to 1:season_nums.

extended
Use full,images to get season posters. Can be min, images, full (default), full,images.

dropunaired
If TRUE (default), episodes which have not aired yet are dropped.

Value

A data.frame containing episode details

Details

This is basically just an extension of trakt.seasons.season, which is used in this function to collect all the episode data. If you only want the episode data for a single season anyway, trakt.seasons.season is recommended, yet this function makes some additions. The use case of this function is to quickly gather episode data of all seasons of a show.

Note

This function is mainly for convenience.

See also

Examples

# NOT RUN {
get_trakt_credentials() # Set required API data/headers
# Manually specifiy seasons
breakingbad.episodes <- trakt.get_all_episodes("breaking-bad", season_nums = c(1, 2, 3, 4, 5))
# Get all the seasons
breakingbad.episodes <- trakt.get_all_episodes("breaking-bad")
# Get first to 3rd season
breakingbad.episodes <- trakt.get_all_episodes("breaking-bad", season_nums = 3)
# }