Search for a show via text query or id

Search for a show or movie with a keyword (e.g. "Breaking Bad") and receive basic info of the first search result. It's main use is to retrieve the ids or proper show/movie title for further use, as well as receiving a quick overview of a show/movie.

trakt.search(query, type = "movie", years = NULL, n_results = 1L,
  extended = c("min", "full"))

trakt.search.byid(id, id_type = c("trakt", "imdb", "tmdb", "tvdb"),
  type = "movie", n_results = 1L, extended = c("min", "full"))

Arguments

query
character(1): The keyword used for the search, e.g. "game of thrones".

type
character(1) ["movie"]: The type of data you're looking for. One of show, movie, episode, person or list or a character vector with those elements, e.g. c("show", "movie"). Note that not every combination is reasonably combinable, e.g. c("movie", "list"). Use separate function calls in that case.

years
integer(1): 4-digit year or range, e.g. 2010-2020. Can also be an integer vector of length two which will be coerced appropriately.

n_results
integer(1) [1]: How many results to return.

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

id
character(1): The id used for the search, e.g. 14701 for a Trakt ID.

id_type
character(1) ["trakt"]: The type of id. One of trakt, imdb, tmdb, tvdb.

Value

A tibble() containing a n_result result. If no results are found, the tibble has 0 rows.

Details

The amount of information returned is equal to .summary API methods and in turn depends on the value of extended. See also the API reference here for which fields of the item metadata are searched by default.

See also

Examples

# A show
trakt.search("Breaking Bad", type = "show", n_results = 3)

#> # A tibble: 3 x 9
#>   type   score title           year trakt  slug          tvdb  imdb        tmdb 
#>   <chr>  <dbl> <chr>          <int> <chr>  <chr>         <chr> <chr>       <chr>
#> 1 show  1000   Breaking Bad    2008 1388   breaking-bad  81189 tt0903747   1396 
#> 2 show   395.  Breaking Bad:…  2018 140684 breaking-bad… 3564… https://ww… <NA> 
#> 3 show    12.0 Breaking In     2011 34392  breaking-in   2067… tt1630574   34541

# A show by its trakt id, and now with more information
trakt.search.byid(1388, "trakt", type = "show", extended = "full")

#> # A tibble: 1 x 29
#>   type  score title  year overview first_aired         runtime certification
#>   <chr> <dbl> <chr> <int> <chr>    <dttm>                <int> <chr>        
#> 1 show   1000 Brea…  2008 Walter … 2008-01-21 02:00:00      45 TV-MA        
#> # … with 21 more variables: network <chr>, country <chr>, trailer <chr>,
#> #   homepage <chr>, status <chr>, rating <dbl>, votes <int>,
#> #   comment_count <int>, updated_at <dttm>, language <chr>,
#> #   available_translations <list>, genres <list>, aired_episodes <int>,
#> #   airs_day <chr>, airs_time <chr>, airs_timezone <chr>, trakt <chr>,
#> #   slug <chr>, tvdb <chr>, imdb <chr>, tmdb <chr>

# A person
trakt.search("J. K. Simmons", type = "person", extended = "full")

#> # A tibble: 1 x 12
#>   type  score name  biography birthday   death birthplace homepage trakt slug 
#>   <chr> <dbl> <chr> <chr>     <date>     <lgl> <chr>      <lgl>    <chr> <chr>
#> 1 pers…  1.54 J.K.… "Jonatha… 1955-01-09 NA    Detroit, … NA       15722 j-k-…
#> # … with 2 more variables: imdb <chr>, tmdb <chr>

# A movie or a show, two of each
trakt.search("Tron", type = c("movie", "show"), n_results = 2)

#> # A tibble: 4 x 9
#>   type   score title            year trakt slug             imdb     tmdb  tvdb 
#>   <chr>  <dbl> <chr>           <int> <chr> <chr>            <chr>    <chr> <chr>
#> 1 movie 1000   Tron             1982 66    tron-1982        tt00848… 97    <NA> 
#> 2 movie  712.  TRON: Legacy     2010 12601 tron-legacy-2010 tt11040… 20526 <NA> 
#> 3 show   591.  Tron: Uprising   2012 34209 tron-uprising    tt18125… 34356 2584…
#> 4 show    10.3 Game of Thrones  2011 1390  game-of-thrones  tt09449… 1399  1213…