Get a user's ratings

Retrieve a user's media ratings

trakt.user.ratings(user = getOption("trakt_username"),
  type = c("movies", "seasons", "shows", "episodes"), rating = NULL,
  extended = c("min", "full"))

Arguments

user
Target username. Defaults to getOption("trakt_username"). Can also be of length greater than 1, in which case the function is called on all user values separately and the result is combined.

type
character(1): Either "shows" or "movies". For season/episode-specific functions, values seasons or episodes are also allowed.

rating
integer(1L): Rating to filter by. Can be 1 through 10, default is NULL

extended
character(1): Either "min" (API default) or "full". The latter returns more variables and should generally only be used if required.

Value

A tibble().

See also

Examples

# \dontrun{
trakt.user.ratings(user = "jemus42")

#> # A tibble: 342 x 9
#>    rated_at            rating type  title      year trakt  slug      imdb  tmdb 
#>    <dttm>               <int> <chr> <chr>     <int> <chr>  <chr>     <chr> <chr>
#>  1 2019-09-29 16:22:36      8 movie Mad Max:…  2015 56360  mad-max-… tt13… 76341
#>  2 2019-09-28 23:34:23      7 movie Tokyo Go…  2003 7877   tokyo-go… tt03… 13398
#>  3 2019-09-27 03:00:59      8 movie Spider-M…  2019 273537 spider-m… tt63… 4296…
#>  4 2019-09-26 03:06:31      8 movie Ghost in…  2004 7057   ghost-in… tt03… 12140
#>  5 2019-09-21 22:58:50      7 movie The Dead…  2019 387368 the-dead… tt86… 5355…
#>  6 2019-09-13 21:38:04      8 movie John Wic…  2019 304278 john-wic… tt61… 4581…
#>  7 2019-08-11 17:32:58      6 movie Hellboy …  2008 6375   hellboy-… tt04… 11253
#>  8 2019-08-11 17:32:58      7 movie Shazam!    2019 184992 shazam-2… tt04… 2879…
#>  9 2019-08-11 17:32:50      7 movie Hellboy    2004 981    hellboy-… tt01… 1487 
#> 10 2019-08-08 21:57:47      7 movie Collater…  2004 998    collater… tt03… 1538 
#> # … with 332 more rows

trakt.user.ratings(user = "sean", type = "movies")

#> # A tibble: 145 x 9
#>    rated_at            rating type  title      year trakt  slug      imdb  tmdb 
#>    <dttm>               <int> <chr> <chr>     <int> <chr>  <chr>     <chr> <chr>
#>  1 2019-09-16 22:10:10      9 movie Under th…  2009 27217  under-th… tt13… 41207
#>  2 2018-07-28 05:41:30      9 movie Watchmen   2009 7729   watchmen… tt04… 13183
#>  3 2018-02-12 20:22:17      9 movie Three Bi…  2017 266092 three-bi… tt50… 3599…
#>  4 2017-12-18 20:58:06      4 movie The Dark…  2017 222255 the-dark… tt16… 3534…
#>  5 2017-12-11 02:14:19      7 movie Kingsman…  2017 223444 kingsman… tt46… 3436…
#>  6 2017-12-11 02:14:09      9 movie American…  2017 260759 american… tt19… 4158…
#>  7 2017-11-16 03:29:00      5 movie Baywatch   2017 220314 baywatch… tt14… 3398…
#>  8 2017-08-02 21:25:33      8 movie Alien: C…  2017 86997  alien-co… tt23… 1268…
#>  9 2017-07-24 22:43:51      9 movie Baby Dri…  2017 219867 baby-dri… tt38… 3394…
#> 10 2017-07-21 04:43:03      8 movie Dunkirk    2017 224641 dunkirk-… tt50… 3747…
#> # … with 135 more rows

# }