Set the required trakt.tv API credentials

trakt_credentials searches for your credentials and stores them in the appropriate options() variables of the same name. To make this work automatically, place your key as environment variables in ~/.Renviron (see Details). Arguments to this function take precedence over any key file. To make API functions work, you do not have to use this function unless you want to supply your own client ID, which is recommended for larger data collection projects.

trakt_credentials(username, client.id, silent = TRUE)

Arguments

username
character(1): Explicitly set your trakt.tv username (optional).

client.id
character(1): Explicitly set your API client ID (required for API interaciton).

silent
logical(1) [TRUE]: No messages are printed showing you the API information. Mostly for debug purposes.

Value

Nothing. Only messages.

Details

Set appropriate values in your ~/.Renviron like this:

# tRakt
trakt_username=jemus42
trakt_client_id=12[...]f2

Note

Please note that no OAuth2 methods are supported yet, only client ID really matters.

See also

Examples

# Use a values set in ~/.Renviron in an R session:
# (This is automatically executed when calling library(tRakt))
trakt_credentials()

# Explicitly set values in an R session, overriding .Renviron values if present
trakt_credentials(
  username = "sean",
  client.id = "12fc1de7671c7f2fb4a8ac08ba7c9f45b447f4d5bad5e11e3490823d629afdf2",
  silent = FALSE
)

#> Your trakt.tv username is set to sean

#> Your client ID is set to 12fc1de7671c7f2fb4a8ac08ba7c9f45b447f4d5bad5e11e3490823d629afdf2